% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbAgStatRec.camp64.R
\name{AbAgStatRec.camp64}
\alias{AbAgStatRec.camp64}
\title{Calcula las abundancias estratificadas por edad y hora en cada rectángulo ICES de la zona considerada}
\usage{
AbAgStatRec.camp64(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  plus = 8,
  mediahora = 2,
  cor.time = TRUE,
  AltAlk = NA,
  incl2 = FALSE,
  DatGraf = FALSE,
  plotrix = TRUE,
  es = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camp}{Campaña de la que se extraen los datos un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa" (proporciona los datos para Medits pero no saca mapas y no tiene sentido ICES rectangulos)}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{mediahora}{Pondera a una hora si se solicita por defecto en 2 suponiendo lances de media hora}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{AltAlk}{ALK alternativa tomada de un fichero de edad del Camp edadXYY.dbf sin ruta ni extensión}

\item{incl2}{Si T incluye los datos de los lances extra, tener en cuenta que no son lances oficiales para calcular los índices de abundancia}

\item{DatGraf}{Si F saca tabla con abundancias en cada rectángulo ICES muestreado en la campaña, si T saca los datos para sacar gráfica con MapIStatRec}

\item{es}{si T saca para etiquetas en español (Ed para edad), si FALSE en inglés (Ag para edad)}
}
\description{
Función de resultados: abundancias estratificadas por edad en cada rectángulo ICES dentro del área de la cqampaña a partir de los datos del camp.
}
\examples{
AbAgStatRec.camp64(1,74,"N21","Cant","local",plus=3,DatGraf = FALSE)
AbAgStatRec.camp64(1,43,"P10","Porc","local",DatGraf = T)
}
\seealso{
Other edades: 
\code{\link{GetAlk.camp64}()},
\code{\link{bubbage.camp64}()},
\code{\link{datagegr.camp64}()},
\code{\link{edadsect.camp64}()},
\code{\link{edadstr.camp64}()},
\code{\link{grafedtal.camp64}()},
\code{\link{grafedtal.camps64}()},
\code{\link{logabage.camp64}()},
\code{\link{maphistage64}()}
}
\concept{edades}
