% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbrvEsp64.R
\name{AbrvEsp64}
\alias{AbrvEsp64}
\title{Crea abreviaturas de especies a partir del nombre científico, género+especie}
\usage{
AbrvEsp64(x, nGen = 3, nEsp = 4, sep = "_", unique = TRUE)
}
\arguments{
\item{x}{Vector con la serie de nombres científicos a abreviar}

\item{nGen}{Número de caracteres del género a utilizar}

\item{nEsp}{número de catacteres del nombre de especie a utilizar}

\item{sep}{Caracter para separar Género de especie}

\item{unique}{Evitar duplicados en los códigos, no tiene mucho sentido pero se puede dejar como falso y tener repetidos los códigos con unique=FALSE}
}
\value{
Devuelve un data.frame con dos columnas, nombres originales $names y los codigos $codes
}
\description{
Transforma series de nombres científicos de especies en abreviaturas de 9 caracteres, se puede elegir el número de
caracteres y el caracter de separación. Ignora borrando el sp. los nombres genéricos de especies con sp., no evita
los spp. cuidado.
Avisa si hay códigos duplicados en el resultado de los códigos solicitados, para evitar crear códigos repetidos para distintas especies haciend uso de base::make.unique
}
\examples{
AbrvEsp64(c("Merluccius merluccius","Merlangius merlangus"))
\dontrun{
AbrvEsp64(buscaesp64(1,50,"cant"),1,8,".")
}
}
\seealso{
Other datos_especies: 
\code{\link{BuscaAphia64}()},
\code{\link{buscacod64}()},
\code{\link{buscaesp64}()},
\code{\link{camptoyear}()}
}
\concept{datos_especies}
