% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agradec.R
\name{agradec}
\alias{agradec}
\title{Transforma grados decimales a grados y minutos decimales
Convierte valores expresados en grados decimales a un formato donde la parte
entera son los grados y la parte decimal representa los minutos decimales
(sin segundos). Por ejemplo, 43.968 se convierte en 43 grados y 0.968*60 minutos.
#' @param x Vector numérico con grados decimales.
#' @examples
#' agradec(43.968)
#' @family Conversión unidades
#' @export}
\usage{
agradec(x)
}
\description{
Transforma grados decimales a grados y minutos decimales
Convierte valores expresados en grados decimales a un formato donde la parte
entera son los grados y la parte decimal representa los minutos decimales
(sin segundos). Por ejemplo, 43.968 se convierte en 43 grados y 0.968*60 minutos.
#' @param x Vector numérico con grados decimales.
#' @examples
#' agradec(43.968)
#' @family Conversión unidades
#' @export
}
