% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALKs64.dns.camp.R
\name{ALKs64.dns.camp}
\alias{ALKs64.dns.camp}
\title{Datos de claves talla-edad en campañas (CampR64)}
\usage{
ALKs64.dns.camp(
  zona = c("cant", "porc", "arsa", "medi"),
  camp = "",
  dns = c("local", "serv")
)
}
\arguments{
\item{zona}{Zona/origen de los datos: Cantábrico "cant", Porcupine "porc",
Golfo de Cádiz "arsa", Mediterráneo "medi"}

\item{camp}{Filtro de campañas: prefijo que sigue a "EDAD" en el nombre del
fichero (p.ej. "P" Porcupine, "N" demersales, "1"/"2" Arsa, "M" Medits...).
Si es "" (por defecto) toma todos los ficheros EDAD* del directorio}

\item{dns}{Origen de las rutas: "local" o "serv"}
}
\value{
Tabla (especies x campañas) con el nº de individuos con datos de edad
}
\description{
Función de acceso a datos. Localiza los ficheros EDAD*.dbf presentes en el
directorio que corresponde a la zona/origen y resume el número de individuos
con datos de edad por campaña y especie. Sustituye al acceso vía RODBC de la
versión de 32 bits (ALKs.dns.camp) por lectura directa de DBF con el lector
unificado read_dbf_simple().
}
\examples{
\dontrun{
ALKs64.dns.camp("porc", "P")
ALKs64.dns.camp("cant", "N")
}
}
\seealso{
\link{readCampDBF}, \link{buscaesp64}

Other edades: 
\code{\link{AbAgStatRec.camp64}()},
\code{\link{GetAlk.camp64}()},
\code{\link{bubbage64.camp}()},
\code{\link{datagegr.camp64}()},
\code{\link{edadsect.camp64}()},
\code{\link{edadstr.camp64}()},
\code{\link{grafedtal.camps64}()},
\code{\link{logabage64.camp}()},
\code{\link{maphistage64}()}
}
\concept{edades}
