% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armap.tot64.r
\name{armap.tot64}
\alias{armap.tot64}
\title{Mapa ampliado de Porcupine con las estaciones realizadas}
\usage{
armap.tot64(
  camp,
  zona = "porc",
  dns = c("local", "serv"),
  lwdl = 1,
  col = 2,
  argr = 2,
  arrow = FALSE,
  leg = FALSE,
  es = FALSE,
  bw = TRUE,
  noval = FALSE,
  CTDs = FALSE,
  strat = FALSE,
  new.dev = TRUE
)
}
\arguments{
\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant" (ARSA para sacar datos al IBTS, no gráficos)}

\item{lwdl}{Anchura de las líneas}

\item{col}{Color de los puntos que representan estaciones}

\item{argr}{Grosor de las flechas de recorrido}

\item{arrow}{Si T añade flechas de reccorrido}

\item{leg}{Si T añade leyenda}

\item{es}{Si T nombres en castellano, si F nombres en inglés}

\item{bw}{Si T mapa en blanco y negro, si F colorea los estratos y sectores}

\item{noval}{Si T añade las estaciones nulas}

\item{CTDs}{Si T añade las estaciones de CTD realizadas}

\item{strat}{strat Si T represent los estratos}
}
\value{
Saca mapa con el desarrollo de la campaña con la zona y estratificación incluyendo tierra (Porcupine)
}
\description{
Crea un mapa de Porcupine con las estaciones realizadas(al igua que "armap.camp") pero con la referencia de Irlanda para su mejor ubicación.
}
\examples{
armap.tot("P14",dns="Porc",noval=TRUE,CTDs=FALSE,bw=TRUE,strat=FALSE,leg=TRUE)
armap.tot("N14",dns="Cant",noval=TRUE,CTDs=FALSE,bw=FALSE,leg=TRUE)
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp64}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{maparea64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()},
\code{\link{maphistal64}()}
}
\concept{mapas}
\concept{resumen general}
