% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/As1.R
\docType{data}
\name{As1}
\alias{As1}
\title{Serie histórica Golfo de Cádiz – campaña AS1}
\format{
An object of class \code{character} of length 31.
}
\usage{
As1
}
\description{
Vector con los códigos de campaña AS1 en el Golfo de Cádiz.
Incluye todos los años disponibles en la serie histórica.
}
\seealso{
Other series históricas: 
\code{\link{As2}},
\code{\link{Psh}}
}
\concept{series históricas}
\keyword{datasets}
