% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/As2.R
\docType{data}
\name{As2}
\alias{As2}
\title{Serie histórica Golfo de Cádiz – campaña AS1}
\format{
An object of class \code{character} of length 28.
}
\usage{
As2
}
\description{
Vector con los códigos de campaña AS1 en el Golfo de Cádiz.
Incluye todos los años disponibles en la serie histórica.
}
\seealso{
Other series históricas: 
\code{\link{As1}},
\code{\link{Psh}}
}
\concept{series históricas}
\keyword{datasets}
