% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom64.r
\name{binom64}
\alias{binom64}
\title{Distribucion binomial: exitos necesarios para ser significativo}
\usage{
binom64(n, p.teorica = 0.5)
}
\arguments{
\item{n}{Número de eventos a tener en cuenta}

\item{p.teorica}{probabilidad teórica de la distribución que se quiere comprobar.}
}
\value{
Devuelve una matriz con tres columnas: una la probabilidad en función del número de casos que se consideren como un éxito, y luego los intervalos de confianza inferior y superior, a partir de que p baja de 1 y llega a .05 se puede considerar que el resultado no se puede considerar debido al azar
}
\description{
Funcion para la MSFD y calculo de numero de especies que hacen que unos resultados esten fuera de lo que cabria esperar al azar para ver mejoras en la biodiversidad por especie.
}
\examples{
#binom(24)
}
