% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbage.camp64.r
\name{bubbage.camp64}
\alias{bubbage.camp64}
\title{Crea gráficos de burbujas de la abundancia sqrt(num)}
\usage{
bubbage.camp64(
  gr,
  esp,
  camps,
  zona = "porc",
  dns = "local",
  plus = 8,
  recr = 0,
  cor.time = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camps}{Serie historica de campañas de la que se extraen los datos. Todas las campañas han de tener ALKs para las especie en cuestión}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{recr}{edad de reclutamiento para especies con reclutamiento en edad 1 o mayor}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}
}
\value{
Saca un lattice de descenso de abundancia logarítmica con la edad con dato de la pendiente para cada año. También saca una matriz de abundancias por años x edad columnas x filas
}
\description{
Tipico bubleplot de análisis de VPA y evaluación de poblaciones con gráfico por edad y año
}
\examples{
bubbage.camp(1,43,Nsh[1:29],"Cant",8,0)
}
\seealso{
Other edades: 
\code{\link{GetAlk.camp64}()},
\code{\link{edadsect.camp64}()},
\code{\link{edadstr.camp64}()},
\code{\link{grafedtal.camp64}()},
\code{\link{grafedtal.camps64}()}
}
\concept{edades}
