% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscaesp64.R
\name{buscaesp64}
\alias{buscaesp64}
\title{Función de búsqueda del nombre de una especie}
\usage{
buscaesp64(gr, esp, zona, dns = c("local", "serv"), id = "l")
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{zona}{Zona de la que se leen las especies. Determina la estructura de la tabla de especies, que difiere entre ARSA y el resto (Porcupine/Cantábrico): "cant", "porc", "arsa", "medi"}

\item{dns}{Origen de las bases de datos: "local" en el ordenador, "serv" en el servidor}

\item{id}{Salida como: "l" Nombre científico en latín, "e" nombre en español, "i" nombre en inglés "a" código APHIA}
}
\description{
Busca el nombre de una especie a partir del código del grupo y especie.Puede dar el nombre científico, común o en inglés
}
\examples{
\dontrun{
buscaesp64(1,50,"cant","local")
}
}
\seealso{
Other datos_especies: 
\code{\link{AbrvEsp64}()},
\code{\link{BuscaAphia64}()},
\code{\link{buscacod64}()},
\code{\link{camptoyear}()}
}
\concept{datos_especies}
