% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAMPtoHLnw64.R
\name{CAMPtoHLnw64}
\alias{CAMPtoHLnw64}
\title{Exporta datos de formato CAMP a formato DATRAs HL. Depende de que los códigos Aphia estén correctos en especies.dbf da error si son incompletos}
\usage{
CAMPtoHLnw64(
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  inclSpecie = FALSE,
  quart = TRUE,
  incl2 = FALSE,
  export = FALSE
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa"}

\item{dns}{elige de dónde se toman los datos, del ordenador ("local")  o del servidor ("serv") si está configurado}

\item{inclSpecie}{si T incluye el nombre de la especie y el Código, si no sólo el Aphia}

\item{quart}{si F deja en cada lance el valor del trimestre en que se realizó el lance, si T se deja el que tiene la campaña por defecto, 1 para Arsa 1Q, 3 para Porcupine y 4 para Arsa 4Q y Demersales Northern Shelf}

\item{incl2}{Si F deja fuera los lances especiales que actualmente no se transmiten a DATRAS, si T los incluye}

\item{export}{Si T crea un fichero csv con todos los datos corregidos (APHIAs) en el directorio CAMP donde está el especies.dbf este es importable al especies.dbf con un append from deli with, quitando todos los peces grupo="1"}
}
\value{
Devuelve un data.table con datos de cada especie en el formato HL de DATRAS. DATRAS requiere que los datos no tengan cabecera y el trimestre sea el que corresponde a la campaña, además de no tener "". Por ello se debe pasar a fichero con la orden: write.table(CAMPtoHH(Xyy,dns),"nombrearchivo.csv",sep=",",quote=F,col.names=F,row.names=F))
}
\description{
Función de Salida de datos a DATRAS:
Extrae las características de las capturas por lance para una campaña desde el fichero NTALLxxx.DBF y los transforma en formato DATRAS HL. De momento sólo funciona con peces y en el SPNGFS y SPPORC (Para completar crustáceos y moluscos hay que añadir los AphiaID, y para ARSA añadirlos al especies de ARSA)
}
\examples{
# CAMPtoHLnw("P14","porc","local")
}
