% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camptoyear.R
\name{camptoyear}
\alias{camptoyear}
\title{Transforma series de nombres de campaña en años}
\usage{
camptoyear(x)
}
\arguments{
\item{x}{Vector con la serie de nombres de campaña a transformar a años}
}
\description{
Transforma series de nombres de campañas en formato Camp XYY a años, si se incluyen códigos de 3 caracteres que no corresponden a caracter número número, devuelve 0, si es
}
\examples{
camptoyear(Nsh)
}
\seealso{
Other datos_especies: 
\code{\link{BuscaAphia64}()},
\code{\link{buscacod64}()},
\code{\link{buscaesp64}()}
}
\concept{datos_especies}
