% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CephDataByH64.R
\name{CephDataByH64}
\alias{CephDataByH64}
\title{Crea la salida para el CPUE por lance del Data call de cefalópodos}
\usage{
CephDataByH64(
  gr = 3,
  esp,
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  quarter = FALSE,
  incl2 = FALSE,
  incl0 = FALSE
)
}
\arguments{
\item{gr}{Grupo SIEMPRE 3 da error si no}

\item{esp}{especie, se pueden incluir varias especies y da la suma de los datos de varias especies, es útil para sacar los datos de un género o una familia.}

\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant, Golfo de Cádiz "Arsa"}

\item{quarter}{si F deja en cada lance el valor del trimestre en que se realizó el lance, si T se deja el que tiene la campaña por defecto, 1 para Arsa 1Q, 3 para Porcupine y 4 para Arsa 4Q y Demersales Northern Shelf}

\item{incl2}{Si F deja fuera los lances especiales que actualmente no se transmiten a DATRAS, si T los incluye}

\item{incl0}{si T se incluyen los lances sin captura de la especie, si F sólo salen los lances con capturas}
}
\value{
Devuelve un data.table con datos de cada especie en el formato HL de DATRAS. DATRAS requiere que los datos no tengan cabecera y el trimestre sea el que corresponde a la campaña, además de no tener "". Por ello se debe pasar a fichero con la orden: write.table(CAMPtoHH(Xyy,dns),"nombrearchivo.csv",sep=",",quote=F,col.names=F,row.names=F))
}
\description{
Función de Salida de datos a DATRAS:
Extrae las características de las capturas por lance para una campaña desde el fichero NTALLxxx.DBF y los transforma en formato DATRAS HL. De momento sólo funciona con peces y en el SPNGFS y SPPORC (Para completar crustáceos y moluscos hay que añadir los AphiaID, y para ARSA añadirlos al especies de ARSA)
}
\examples{
# CephDataByH(3,10,"P14","Porc")
}
