% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurarCampR64.R
\name{configurarCampR64}
\alias{configurarCampR64}
\title{Configurar rutas de CampR64}
\usage{
configurarCampR64(env = "local", especies, cant, porc, arsa, medi)
}
\arguments{
\item{env}{"local" o "server"}

\item{especies}{Ruta al directorio con ESPECIES.DBF}

\item{cant}{Ruta a los datos de Cantábrico}

\item{porc}{Ruta a los datos de Porcupine}

\item{arsa}{Ruta a los datos de ARSA}

\item{medi}{Ruta a los datos de MEDITS}
}
\description{
Permite establecer las rutas locales o de servidor para los datos
de campañas y el fichero ESPECIES.DBF. La configuración se guarda
en \verb{~/.CampR64/configRoots_user.R}.
}
