% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.camp64.R
\name{CV.camp64}
\alias{CV.camp64}
\title{Abundancias, SE y coeficientes de variación paramétricos por subestrato, estrato y total}
\usage{
CV.camp64(
  gr,
  esp,
  camp,
  zona,
  dns = "local",
  ind = "p",
  cor.time = TRUE,
  kg = TRUE,
  Nas = FALSE,
  excl.sect = NA,
  verbose = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros. 9 incluye todos los grupos a excepción del 6}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que sacar los dato un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{ind}{Dato a procesar, si (p)eso o (n)umero}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{kg}{Si T saca los resultados en Kg, si F en gramos}

\item{Nas}{Si F el error estándar de los subestratos y estratos con un solo lance son NA. Si T utiliza el valor del estrato y evita los NAs}

\item{excl.sect}{Excluye los sectores o subsectores dados como caracteres}

\item{verbose}{si T avisa de que distintas especies pueden pertenecer a grupos distintos y hay que valorar el sendio que tiene mezclarlas}
}
\value{
Devuelve una lista con valorees de media, SE y CV para los subestratos, estratos y total de la campaña
}
\description{
Función de resultados
}
\seealso{
{\link{CV.camps}}, {\link{CV.bt.camp}}
}
