% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.camps64.R
\name{CV.camps64}
\alias{CV.camps64}
\title{Abundancias en número y peso totales de la especie para varias campañas}
\usage{
CV.camps64(
  gr,
  esp,
  camps,
  zona,
  dns = c("local", "serv"),
  cor.time = TRUE,
  kg = TRUE,
  dec = 2,
  excl.sect = NA,
  verbose = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros. 9 incluye todos los grupos a excepción del 6}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{Campañas de las que se solicitan los resultados un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cadiz "arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{dns}{Elige de dónde se leen los datos si del ordenador "local" o del servidor "serv"}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{kg}{Si T saca los resultados en Kg, si F en gramos}

\item{dec}{Número de decimales de los resultados}

\item{excl.sect}{Excluye los sectores o subsectores dados como caracteres}

\item{verbose}{si T avisa que disintas especies pueden pertenercer a grupos distintos y hay que valorar si se pueden juntar}
}
\value{
Devuelve una lista con valorees de media, SE y CV para los subestratos, estratos y total de la campaña
}
\description{
Función de resultados
}
\seealso{
ListFauna.camp
}
