% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datab64.R
\name{datab64}
\alias{datab64}
\title{Datos estratificados por sector para una especie para grupos de trabajo}
\usage{
datab64(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  cor.time = TRUE,
  Nas = FALSE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{Nas}{Permite calcular los errores estándar aunque sólo haya un lance en algún estrato (haciendo varianza =0 en ese estrato, incorrecto pero da una idea cuando sólo un estrato entre varios tiene sólo un lance)}
}
\value{
Devuelve un número con nombres organizado en dos líneas (biomasa y número) en columnas por sectores geográficos segun los definidos en el Camp e información por columnas abundancia estratificada media por estrato (avgestr) y error estándar (SEestr) y totales (avgsect, SEsect). Preparado para pegarlo de año en año en los ficheros excel de abundancia en grupo de trabajo
}
\description{
Salida de datos estratificados por sector a csv para rellenar los informes de grupo de trabajo, filas con datos ab estratificada (Biomasa y N) y error estándar por sector. Da la abundancia y error estándar en cada sector (geográfico definido en el Camp)
}
\seealso{
{\link{databICES} \link{databICESdiv} \link{databEstr}}
}
