% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databEstr64.R
\name{databEstr64}
\alias{databEstr64}
\title{Salida con abundancias estratificadas por estrato de profundidad}
\usage{
databEstr64(
  gr,
  esp,
  camp,
  zona = c("cant"),
  dns = c("local", "serv"),
  cor.time = TRUE,
  Nas = FALSE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa"}

\item{dns}{Elige si los datos se toman del ordenador ("local") o del servidor ("serv")}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{Nas}{permite calcular los errores estándar aunque sólo haya un lance en algún estrato (haciendo varianza =0 en ese estrato, incorrecto pero da una idea cuando sólo un estrato entre varios tiene sólo un lance)}
}
\value{
Devuelve un número con nombres organizado en dos líneas (biomasa y número) en columnas por estratos de profundidad segun definidos en el Camp e información por columnas abundancia estratificada media por estrato (avgestr) y error estándar (SEestr) y totales (avgsect, SEsect). Preparado para pegarlo de año en año en los ficheros excel de abundancia en grupo de trabajo
}
\description{
Salida de datos a csv para rellenar los informes de grupo de trabajo por estrato, filas con datos ab estratificada (Biomasa y N) y error estándar por estrato. Da la abundancia y error estándar en cada estrato de profundidad (definido en el Camp)
}
\examples{
databEstr(1,51,"N05","Cant")
}
\seealso{
{\link{databICES64} \link{databICESdiv64} \link{datab64}}
}
