% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databICESDiv64.R
\name{databICESdiv}
\alias{databICESdiv}
\title{Datos de abundancia por división ICES para una especie resúmen para grupos de trabajo}
\usage{
databICESdiv(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  cor.time = TRUE,
  Nas = FALSE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa"}

\item{dns}{Elige si los datos se toman del ordenador ("local") o del servidor ("serv")}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{Nas}{Permite calcular los errores estándar aunque sólo haya un lance en algún estrato (haciendo varianza =0 en ese estrato, incorrecto pero da una idea cuando sólo un estrato entre varios tiene sólo un lance)}
}
\value{
Devuelve un número con nombres organizado en dos líneas (biomasa y número) en columnas por subdivisiones ICES por columnas abundancia estratificada media por XIa, 8.cE, 8.cW
}
\description{
Salida de datos a csv para rellenar los informes de grupo de trabajo, filas con datos ab estratificada (Biomasa y N) y error estándar por subdivisión ICES función para Demersales Norte (saca 9.a, 8.c y total)
}
\seealso{
{\link{databICES} \link{databEstr} \link{datab}}
}
