% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagegr.camp64.R
\name{datagegr.camp64}
\alias{datagegr.camp64}
\title{Extrae los datos del FAUNA de una especie en concreto en una edad.}
\usage{
datagegr.camp64(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  plus = 8,
  excl.sect = NA,
  cor.time = TRUE,
  n.ots = FALSE,
  AltAlk = NA,
  incl2 = TRUE,
  mediahora = 1
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustaceos: Sólo existen datos de edad para algunos peces y la cigala.}

\item{esp}{Codigo de la especie numerico o caracter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña o campañas a representar en el mapa}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa".}

\item{dns}{elige si se trabja con archivos del ordenador ("local") o del servidor ("serv")}

\item{plus}{Edad plus: incluir la edad considerada como plus, solo afecta si se pide la edad solicitada que une todas las edades mayores}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos, NA no excluye ninguno}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{n.ots}{Valor interno para decir que no se saque el número de otolitos en la clave sino las proporciones}

\item{AltAlk}{Clave talla edad alternativa sin ruta ni extensión, NA por defecto usa la clave de la campaña edadXYY.dbf}

\item{incl2}{si T coge los datos de lances especiales, si F los excluye}

\item{mediahora}{Valor para obtener abundancias por hora si media hora es mayor}
}
\value{
Devuelve un data.frame con campos lance, lat, long y abundancia por edad 0,1,2...Plus de edad, lance, peso número subestrato...
}
\description{
Función de acceso a datos: saca las abundancias por edad y lance de la especie
}
\seealso{
Other edades: 
\code{\link{AbAgStatRec.camp64}()},
\code{\link{GetAlk.camp64}()},
\code{\link{bubbage.camp64}()},
\code{\link{edadsect.camp64}()},
\code{\link{edadstr.camp64}()},
\code{\link{grafedtal.camp64}()},
\code{\link{grafedtal.camps64}()},
\code{\link{logabage.camp64}()},
\code{\link{maphistage64}()}
}
\concept{edades}
