% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datCatches.camp64.R
\name{datCatches.camp64}
\alias{datCatches.camp64}
\title{Crea datos de capturas en formato para evaluaciones tipo Capros o modelos bayesianos (completar con datTalCatch.camp)}
\usage{
datCatches.camp64(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = "local",
  cor.time = TRUE,
  incl2 = FALSE
)
}
\arguments{
\item{gr}{Grupo, 1 peces, 2 crustaceos...}

\item{esp}{Código de especie}

\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)gr Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{cor.time}{Corrección del tiempo de arrastre al calcular las abundancias (mantener en T, \emph{da datos por media hora de lance})}

\item{incl2}{Si T se incluyen los lances extra no incluidos para las abundancias o biomasas estratificadas}
}
\value{
A data.table con el formato de datos para otras especies con formato de geográfico por lances
}
\description{
Function para geographical - bayesian models
}
\examples{
# datCatches.camp(1,18,"N23","Cant")
# datCatches.camp(1,18,"P23","Porc")
}
