% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datgr.camp64.r
\name{datgr.camp64}
\alias{datgr.camp64}
\title{Datos de biomasa y abundancia para una especie y campaña}
\usage{
datgr.camp64(
  gr,
  esp,
  camp,
  zona,
  dns = c("local", "serv"),
  cor.time = TRUE,
  incl2 = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros. 9 incluye todos los grupos a excepción del 6}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que se extraen los datos: año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{cor.time}{Si T corrige abundancias con la duración del lance para llevarlo a 30 minutos}

\item{incl2}{Si F no tiene en cuenta los lances especiales, si T si los tiene en cuenta, pero da problemas por que no puede calcular las abundancias estratificadas}
}
\value{
Devuelve un data.frame con información del lance, latitud, longitud, profundidad, peso (gramos) y numero
}
\description{
Función de acceso a datos:
Extrae los datos de peso y número de una especie o conjunto de especies a partir de las faunísticas de una campaña
}
\examples{
datgr.camp(1,50,"P10",zona="porc",dns="local",cor.time=TRUE,incl2=FALSE)
}
