% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dathidro.camp64.R
\name{dathidro.camp64}
\alias{dathidro.camp64}
\title{Datos hidrográficos de una campaña}
\usage{
dathidro.camp64(
  camp,
  zona,
  dns = c("local", "serv"),
  year = TRUE,
  quarter = TRUE
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: "cant", "porc", "arsa", "medi"}

\item{dns}{Elige de dónde se leen los datos, "local" para el ordenador, "serv" para el servidor}

\item{year}{si TRUE incluye una columna con el año al final de los datos}

\item{quarter}{si TRUE incluye una columna con el trimestre de los datos teniendo en cuenta la fecha del lance}
}
\value{
Devuelve un data.frame con datos de cada estación hidrográfica y la correspondencia a los lances del fichero lance
}
\description{
Datos hidrográficos de una campaña
}
\examples{
  print(dathidro.camp64("N25",zona="cant",dns=c("local")))
}
\seealso{
{\link{MapLansGPS}}
}
