% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.camp64.R
\name{datos.camp64}
\alias{datos.camp64}
\title{Datos de abundancia y biomasa de una especie}
\usage{
datos.camp64(
  gr,
  esp,
  camp,
  zona,
  dns = "local",
  cor.time = TRUE,
  kg = TRUE,
  verbose = TRUE,
  incl2 = FALSE
)
}
\arguments{
\item{gr}{grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros. 9 incluye todos los grupos a excepción del 6}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{kg}{Si T datos en kgs, si F en gramos}

\item{verbose}{si T avisa de que hay más de una especie y los datos mezclados pueden ser engañosos}
}
\value{
Devuelve un data.frame con los datos de la especie por lance: sector,lance,peso,numero,arsect (área del sector al que pertence el lance)
}
\description{
Función de acceso a datos:
Extrae los datos de abundancia y biomasa estratificados (solo lances con validez 1) de una especie o conjunto de especies a partir de las faunísticas de una campaña. Crea una tabla con información del sector(número) y estrato (letra), lance, peso, número y arsect
Funciones para obtener data.frame de especies concretas a partir de los ficheros del Camp
}
\details{
Saca los datos de los lances estratificados, por ello se produce un error si encuentra un lance con validez 1 y estrato o sector sin información.
}
\examples{
datos.camp(1,50,"P10","Porc",kg=FALSE, cor.time=TRUE)
}
