% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datTalCatch64.camp.R
\name{datTalCatch64.camp}
\alias{datTalCatch64.camp}
\title{Crea datos de capturas en formato para evaluaciones tipo Capros o modelos bayesianos (completar con datCatches.camp)}
\usage{
datTalCatch64.camp(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  sex = TRUE,
  cor.time = TRUE,
  incl2 = FALSE
)
}
\arguments{
\item{gr}{Grupo, 1 peces, 2 crustaceos...}

\item{esp}{Código de especie}

\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{"cant","porc","arsa","medi".}

\item{dns}{"local" (por defecto) o "serv".}

\item{sex}{si T incluye la información por sexos, si F no tiene en cuenta la información por sexos y lo trata todo como indeterminados}

\item{cor.time}{Corrección del tiempo de arrastre al calcular las abundancias (mantener en T, \emph{da datos por media hora de lance})}

\item{incl2}{Si T se incluyen los lances extra no incluidos para las abundancias o biomasas estratificadas}
}
\value{
A data.table con el formato de datos para otras especies con formato de geográfico por lances
}
\description{
Function to create geographical - bayesian models
}
\examples{
# datCatches.camp64(1,18,"N23","Cant")
# datCatches.camp64(1,18,"P23","Porc")
}
