% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denstall.camp64.r
\name{denstall.camp64}
\alias{denstall.camp64}
\title{Función de densidad distribución de tallas}
\usage{
denstall.camp64(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  cor.time = TRUE,
  ti = FALSE,
  bw = TRUE,
  es = TRUE,
  plot = TRUE,
  idi = "l",
  layout = NA,
  years = TRUE,
  cexleg = 1,
  verbose = FALSE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos. 4 equinodermos y 5 invertebrados no se miden}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{dns}{elige de dónde se toman los datos, si del ordenador "local" o del servidor "serv"}

\item{cor.time}{Corrige la abundancia por la duración del lance frente al estándar de la campaña, al ser distribución de frecuencias no afecta a resultados pero se incluye por consistencia con otras fundiones de esta librería}

\item{ti}{Si T añade título al gráfico, el nombre de la especie en latín}

\item{bw}{Gráfico en blanco y negro si T o en color si F}

\item{es}{Si T gráfico en castellano, si F gráfico en inglés}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}

\item{years}{Si T saca los años como nombre de campaña en los paneles lattice de campañas}

\item{cexleg}{Varía el tamaño de letra de los ejes y del número de la leyenda}
}
\description{
Crea gráfico con la función de densidad de la distribución de tallas estratificada para la especie y las campañas seleccionadas
}
\examples{
denstall.camp(gr=1,esp=50,camp=Psh,dns="Porc",years=TRUE)
}
\seealso{
Other Distribuciones de tallas: 
\code{\link{dtallan.camp64}()},
\code{\link{dtalldens.camp64}()}
}
\concept{Distribuciones de tallas}
