% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DpthPrfl64.r
\name{DpthPrfl64}
\alias{DpthPrfl64}
\title{Perfil de distribución por profundidad}
\usage{
DpthPrfl64(
  gr,
  esp,
  camps,
  zona = "porc",
  dns = "local",
  cor.time = TRUE,
  ind = "p",
  es = TRUE,
  ti = TRUE,
  idi = "l",
  xmax = NA,
  nlans = TRUE,
  spl = FALSE,
  brks = "Sturges",
  tabres = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustaceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numerico o caracter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{Campañas de las que se obtiene la distribución de profundidades (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{dns}{Elige de dónde toma los datos si del ordenador "local" o del servidor "serv"}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{ind}{Parámetro a representar, saca los datos en "p"eso o "n"úmero}

\item{es}{Si T gráfico en castellano, si F gráfico en inglés}

\item{ti}{Si T añade título al gráfico, el nombre de la especie en latín}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{xmax}{Valor máximo del eje x}

\item{nlans}{Si T añade el número de lances en cada rango de profundidad}

\item{spl}{Si T incluye una curva spline en el gráfico}

\item{brks}{Especifica los rangos de profundidad:"Sturges" cada 100 metros, "norte" estratificación de Demersales, "porcupine" estratificación de Porcupine, "FD" cada 50 metros}

\item{tabres}{Muestra una tabla resumen del total de lances, media y total de biomasa o número y frecuencia de la especie por estrato según el brks especificado}
}
\description{
Crea un gráfico de perfil de distribución por profundidad de la biomasa o número de una especie o grupo de especies a partir de los datos de peso y número de la faunística tomados en una campaña
}
\examples{
DpthPrfl(1, 50, "N08", "Cant",brks = "norte",tabres=TRUE,ind="p")
DpthPrfl(1,50,"P08","Porc",brks="porcupine",ti=TRUE)
}
\seealso{
{\link{DpthPrflTals}}
}
