% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DpthPrflTals64.r
\name{DpthPrflTals64}
\alias{DpthPrflTals64}
\title{Perfil de distribución por profundidad y tallas}
\usage{
DpthPrflTals64(
  gr,
  esp,
  camps,
  zona = "porc",
  dns = "local",
  tmin = 0,
  tmax = 999,
  cor.time = TRUE,
  incl2 = T,
  ind = "n",
  sex = NA,
  es = TRUE,
  ti = TRUE,
  idi = "l",
  xmax = NA,
  nlans = TRUE,
  spl = FALSE,
  brks = "Sturges",
  tabres = TRUE,
  tit2 = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustaceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numerico o caracter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{Campañas de las que se obtiene la distribución de profundidades (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa"}

\item{dns}{Elige de dónde se cogen los datos, si del ordenador "local" o del servidor "serv" Si está configurado}

\item{tmin}{Talla mínima}

\item{tmax}{Talla máxima}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{ind}{Parámetro a representar, saca los datos en "p"eso o "n"úmero}

\item{sex}{Permite elegir entre machos(1), hembras(2) o indeterminados(3), NA escoge sin tener en cuenta el sexo}

\item{es}{Si T gráfico en castellano, si F gráfico en inglés}

\item{ti}{Si T añade título al gráfico, el nombre de la especie en latín}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{xmax}{Valor máximo del eje x}

\item{nlans}{Si T añade el número de lances en cada rango de profundidad}

\item{spl}{Si T incluye una curva spline en el gráfico}

\item{brks}{Especifica los rangos de profundidad:"Sturges" cada 100 metros, "norte" estratificación de Demersales, "porcupine" estratificación de Porcupine, "FD" cada 50 metros}

\item{tabres}{Muestra una tabla resumen de la media, total de biomasa o número y frecuencia de la especie por estación según el brks especificado}

\item{tit2}{Añade un segundo título al gráfico especificando el rango de tallas}
}
\description{
Crea un gráfico de perfil de distribución por profundidad de la abundancia de una especie o grupo de especies en un rango de tallas específico a partir de los datos de distribución de tallas tomados en una campaña. También puede sacar los datos en peso pero a partir de las relaciones talla-peso, en vez de a partir de las faunísticas como en DpthPrfl
}
\examples{
DpthPrflTals64(1, 50, "N08", "cant","local",10,20,brks = "norte",tabres=TRUE,ind="p")
DpthPrflTals64(1,50,"P08","porc","local",10,20,brks="porcupine")
}
\seealso{
{\link{DpthPrfl}}
}
