% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtallan.camp64.R
\name{dtallan.camp64}
\alias{dtallan.camp64}
\title{Detalle de tallas por lance (CampR64) con correcciones de tiempo y peso}
\usage{
dtallan.camp64(
  grupo,
  especie,
  camp,
  zona,
  dns = c("local", "serv"),
  sexo = TRUE,
  excl.sect = NA,
  cor.time = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{grupo, especie, camp, zona, dns}{Como en el resto del paquete}

\item{sexo}{TRUE → devuelve columna SEXO (1,2,3); FALSE → agrupa sexos}

\item{excl.sect}{Excluir sectores (si procede)}

\item{cor.time}{TRUE → divide por weight.time (histórico)}

\item{verbose}{Mensajes}
}
\value{
data.frame:
\itemize{
\item sexo=TRUE  → lance, sector, talla, sexo, n
\item sexo=FALSE → lance, sector, talla, n
}
}
\description{
Detalle de tallas por lance (CampR64) con correcciones de tiempo y peso
}
