% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtallmean64.serie.R
\name{dtallmean64.serie}
\alias{dtallmean64.serie}
\title{Distribución de tallas media de una serie de campañas}
\usage{
dtallmean64.serie(
  gr,
  esp,
  camps,
  zona = "cant",
  dns = c("local", "serv"),
  excl.sect = NA,
  cor.time = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos, 3 moluscos,
4 equinodermos, 5 invertebrados}

\item{esp}{Código numérico o carácter (3 espacios) de la especie}

\item{camps}{Vector de campañas}

\item{zona}{"cant", "porc", "arsa"}

\item{dns}{"local" o "serv"}

\item{excl.sect}{Sectores a excluir}

\item{cor.time}{Si TRUE corrige por duración del lance}
}
\value{
data.frame con columnas: \code{talla}, \code{numero}
}
\description{
Devuelve la distribución de tallas promediada sobre el conjunto de
campañas indicado en \code{camps}. Útil como insumo para gráficos
comparativos o cálculos posteriores.
}
\examples{
\dontrun{
  dtallmean64.serie(1, 98, Nsh[14:24], "cant")
}
}
\seealso{
\code{\link{dtallbarplot64}}, \code{\link{dtall.camp64}}
}
