% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edadsect.camp64.R
\name{edadsect.camp64}
\alias{edadsect.camp64}
\title{Calcula las abundancias estratificadas por edad}
\usage{
edadsect.camp64(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  plus = 8,
  excl.sect = NA,
  cor.time = TRUE,
  AltAlk = NA
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camp}{Campaña de la que se extraen los datos un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{dns}{Elige de dónde toma los datos, del ordenador "local" o del servidor "serv"}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{excl.sect}{excluye sectores para calcular el resultado por lo sectores individualmente o en conjunto.}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{AltAlk}{ALK alternativa tomada de un fichero de edad del Camp edadXYY.dbf sin ruta ni extensión}
}
\description{
Función de resultados: abundancias estratificadas por edad para cada sector geográfico a partir de los datos del camp.
}
\examples{
edadsect.camp64(1,43,"P01","porc","local",plus=8)

}
\seealso{
Other edades: 
\code{\link{AbAgStatRec.camp64}()},
\code{\link{GetAlk.camp64}()},
\code{\link{bubbage.camp64}()},
\code{\link{datagegr.camp64}()},
\code{\link{edadstr.camp64}()},
\code{\link{grafedtal.camp64}()},
\code{\link{grafedtal.camps64}()},
\code{\link{logabage.camp64}()},
\code{\link{maphistage64}()}
}
\concept{edades}
