% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fauna.camp64.R
\name{Fauna.camp}
\alias{Fauna.camp}
\title{Listado de especies por campaña sin  datos de las especies del grupo, sólo nombres permite añadir los lances especiales}
\usage{
Fauna.camp(
  gr = "1",
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  excl.sect = NA,
  incl2 = TRUE,
  verbose = FALSE,
  getaphia = FALSE
)
}
\arguments{
\item{camp}{Campaña de la que se extrae el listado de especies capturadas: un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa"}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos, NA no excluye ninguno}

\item{incl2}{si T incluye los lances especiales con listado de especies en los lances especiales y en los estándar}

\item{verbose}{si T saca en pantalla, si F no salen en pantalla}

\item{getaphia}{si T busca en internet (debe estar disponible) los códigos AphiaID de las especies en cuestión}
}
\value{
Devuelve un data.frame con el listado de taxones (especies) del grupo gr encontradas en la campaña
}
\description{
Listado de especies por campaña sin  datos de las especies del grupo, sólo nombres permite añadir los lances especiales
}
\details{
Muestra un listado de todas las especies capturadas del grupo gr. se puede utilizar para sacar listado de todos los lances o sólo de los estándares. Si hay especies con código desonocido aparecen al final del listado#' @param gr Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros.
}
\examples{
Fauna.camp(gr=1,camp="N12",dns="Cant",excl.sect=NA,getaphia=TRUE)
}
\seealso{
Other ListadosFauna: 
\code{\link{ListFauna.camp64}()}
}
\concept{ListadosFauna}
