% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_utils.R
\name{get_shape_path}
\alias{get_shape_path}
\title{Ruta a un shape interno de CampR64}
\usage{
get_shape_path(type, file)
}
\arguments{
\item{type}{Tipo: "ices", "cant", "porc", "arsa", "iberia", "bathy"}

\item{file}{Nombre del fichero (con extension .gpkg)}
}
\value{
Ruta completa al fichero
}
\description{
Ruta a un shape interno de CampR64
}
