% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAlk.camp64.R
\name{GetAlk.camp64}
\alias{GetAlk.camp64}
\title{Extrae ALK para una especie de los ficheros del camp

Función de acceso a datos: recupera la clave talla edad de una especie y campaña determinadas}
\usage{
GetAlk.camp64(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = "local",
  plus = 8,
  n.ots = FALSE,
  AltAlk = FALSE,
  keep_sexo = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camp}{Campaña de la que se extraen los datos un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{dns}{Elige el de dónde se toman los datos si del ordenador mismo "local" o del servidor en el IEO "serv"}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{n.ots}{Número de otolitos o proporción? Si T da el número de otolitos}

\item{AltAlk}{ALK alternativa tomada de un fichero edadXYY.dbf sin ruta ni extensión}

\item{keep_sexo}{nantiene la columna de sexo para claves talla edad que usan el sexo por defecto =FALSE}
}
\description{
Extrae ALK para una especie de los ficheros del camp

Función de acceso a datos: recupera la clave talla edad de una especie y campaña determinadas
}
\examples{
GetAlk.camp(1,43,"N93","Cant",8) 
GetAlk.camp(1,45,"P03","Porc",AltAlk="edadXYY") 
}
\seealso{
Other edades: 
\code{\link{bubbage.camp64}()},
\code{\link{edadsect.camp64}()},
\code{\link{edadstr.camp64}()},
\code{\link{grafedtal.camp64}()},
\code{\link{grafedtal.camps64}()}
}
\concept{edades}
