% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICESarea64.R
\name{getICESarea64}
\alias{getICESarea64}
\title{StatRec y División ICES (Area) para una campaña — versión estricta sin helpers
Requisitos (en lances, todo en minúsculas):
latitud_l, latitud_v, longitud_l, longitud_v, prof_l, prof_v, validez
(y con redux=TRUE: lat, long, prof)
Requisitos (en Area): ICESNAME, Area}
\usage{
getICESarea64(
  camp,
  zona,
  dns = c("local", "serv"),
  incl2 = TRUE,
  incl0 = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{dns}{Elige origen datos ordenador "local" o del servidor "serv"}

\item{incl2}{Si TRUE incluye los datos de los lances especiales (2 o mayor)}

\item{incl0}{Si no es TRUE excluye los lances nulos (0)}
}
\description{
StatRec y División ICES (Area) para una campaña — versión estricta sin helpers
Requisitos (en lances, todo en minúsculas):
latitud_l, latitud_v, longitud_l, longitud_v, prof_l, prof_v, validez
(y con redux=TRUE: lat, long, prof)
Requisitos (en Area): ICESNAME, Area
}
