% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradms.R
\name{gradms}
\alias{gradms}
\title{Transforma grados, minutos y segundos a grados decimales}
\usage{
gradms(x)
}
\arguments{
\item{x}{Vector de caracteres con formato \code{"GGG:MM:SS"}.}
}
\description{
Convierte valores expresados como \code{"GGG:MM:SS"} a grados decimales.
Avisa si los minutos o segundos son mayores de 60.
}
\examples{
gradms("43:14:25")
}
\seealso{
Other Conversion unidades: 
\code{\link{gradec}()}
}
\concept{Conversion unidades}
