% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafedtal.camps64.R
\name{grafedtal.camps64}
\alias{grafedtal.camps64}
\title{Histograma de distribución de tallas con edades para varias campañas

Función gráfica a partir de los ficheros del camp. Histograma distribución de tallas estratificada por edad en varias campañas}
\usage{
grafedtal.camps64(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = "local",
  plus = 8,
  cor.time = TRUE,
  excl.sect = NA,
  ti = FALSE,
  leg = TRUE,
  es = TRUE,
  layout = NA,
  idi = "l",
  cexleg = 1,
  clms = 2,
  plot = TRUE,
  ymax = NA,
  out.dat = FALSE,
  years = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camp}{Campañas a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{dns}{elije de dónde se toman los datos, "local" del mismo ordendor, "serv" del servidor en el IEO si está preparado}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{ti}{Si T añade título al gráfico, el nombre de la especie en latín}

\item{leg}{Si T añade leyenda}

\item{es}{Si T gráfico en castellano, si F gráfico en inglés}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{clms}{Número de columnas para ordenar la serie histórica}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{ymax}{Valor máximo del eje y}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{years}{Si T saca años en lugar de nombres de la campaña}
}
\value{
Saca gráfica con distribución de tallas y la distribución de las edades en cada talla. Si out.dat=TRUE saca un data.frame con columnas n(valor del número de la distribución estratifcada para la talla y la edad),talla,edad,camp. Da error si no existe ALK para la especie en la campaña
}
\description{
Histograma de distribución de tallas con edades para varias campañas

Función gráfica a partir de los ficheros del camp. Histograma distribución de tallas estratificada por edad en varias campañas
}
\examples{
grafedtal.camps(1,43,Nsh[22:31],"Cant",plus=7,years=T,es=FALSE)
}
\seealso{
{\link{grafedtal.camp}}

Other edades: 
\code{\link{AbAgStatRec.camp64}()},
\code{\link{GetAlk.camp64}()},
\code{\link{bubbage.camp64}()},
\code{\link{datagegr.camp64}()},
\code{\link{edadsect.camp64}()},
\code{\link{edadstr.camp64}()},
\code{\link{grafedtal.camp64}()},
\code{\link{logabage.camp64}()},
\code{\link{maphistage64}()}
}
\concept{edades}
