% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafhistbox64.comp.r
\name{grafhistbox64.comp}
\alias{grafhistbox64.comp}
\title{Gráficas grafhistbox combinadas biomasa y peso}
\usage{
grafhistbox64.comp(
  gr,
  esp,
  camps,
  zona = "porc",
  dns = c("local", "serv"),
  cor.time = TRUE,
  kg = TRUE,
  ci.lev = 0.8,
  DLS = F,
  idi = "l",
  SE = TRUE,
  es = TRUE,
  sector = NA,
  Nas = FALSE,
  excl.sect = NA,
  ymax = c(NA, NA),
  tline = FALSE,
  years = TRUE,
  ti = TRUE,
  mar = NA,
  nboot = 1000
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros. 9 incluye todos los grupos a excepción del 6}

\item{esp}{Codigo de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{campañas de la serie de datos a representar en el gráfico de abundancias Demersales Nsh, Porcupine Psh, Arsa primavera As1 y Arsa otoño As2}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa"}

\item{dns}{Elige de dónde se toman los datos "local" del mismo ordenador, "serv" del servidor si está configurado {\link{configurarCampR64}}}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{kg}{Si T el gráfico está en kgs, si F en gramos}

\item{ci.lev}{El intervalo de confianza a representar}

\item{DLS}{si T dibuja líneas en los valores medios de los últimos dos años y de los cinco previos}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{SE}{Si T dibuja las cajas representando el error estándar, si F no aparecen la cajas}

\item{es}{Si T ejes y unidades en español, si F en inglés}

\item{sector}{Alternativa a excl.sect para un sólo sector. Si especificado como carácter solo toma el sector elegido}

\item{Nas}{Si F no representa las cajas en los sectores/estratos en los que algún sector o estrato tiene un sólo lance. Si T utiliza el valor del estrato y evita los NAs}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{ymax}{Valor máximo del eje de las y, tiene doble valor para regular biomasa y número}

\item{tline}{Si T dibuja una línea de tendencia a traves de un glm con los datos de abundancia. Gráficos evaluación MSFD.}

\item{years}{Si T saca los años como nombre de campaña en el eje de las equis en vez del nombre de campaña}

\item{ti}{Permíte sólo el nombre de la especie que aparece en el gráfico superior y sale siempre en cursiva.}

\item{mar}{Si se quiere dejar un margen ya establecido hacerlo igual a F}

\item{nboot}{Número de réplicas bootstrap. Reducir (ej. 200) para exploración rápida, 1000 para resultados finales.}

\item{sub}{Añade un subtítulo con el valor que se le ponga si no F}
}
\value{
Crea una gráfica doble de evolución de las abundancias en biomasa y número.
}
\description{
Ver documentación grafhistbox
}
\examples{
grafhistbox64(1,45,Nsh[7:27],"Cant",DLS=T,es=FALSE,years=TRUE,tline=TRUE,ti=TRUE,sub=TRUE)
}
\seealso{
{\link{grafhistbox64}}, {\link{grafhistbox64.comp}}
}
