% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrafMarksGPS64.R
\name{GrafMarksGPS}
\alias{GrafMarksGPS}
\title{Gráficos de dragas o ctds a partir de las marcas del pescawin}
\usage{
GrafMarksGPS(
  camp,
  zona = "cant",
  dns = "local",
  event = "Draga",
  add = FALSE,
  ti = TRUE,
  years = TRUE,
  pch = NA,
  ptbg = NA,
  xlims = NA,
  ylims = NA,
  cuadr = TRUE,
  es = T,
  places = TRUE,
  bw = FALSE
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa"}

\item{dns}{Elige de dónde se toman los datos, "local" del propio ordenador, "serv" del servidor si está configurado}

\item{event}{Elige el evento que se representa, dragas o CTDs}

\item{add}{Si T añade los puntos al gráfico previo, si F saca el gráfico nuevo}

\item{ti}{Si T incluye un título con el nombre de la campaña}

\item{years}{Si T cambia el código de campaña por el año de la campaña si tiene formato camp (ver parámetro camp)}

\item{ptbg}{Definido como NA, si tiene valor 1:4 o colores definidos en R cambia el color del punto de los datos.}

\item{xlims}{Define los limites longitudinales del mapa, si se deja en NA toma los límites de long del área definida en la campaña}

\item{ylims}{Define los limites latitudinales del mapa, si se deja en NA toma los límites de lat del área definida en la campaña}

\item{cuadr}{Si T por defecto, incluye cuadrículas de 5x5 millas, si F no incluye las líneas}

\item{es}{Si T por defecto los carteles salen en español, en caso contrario en inglés}

\item{places}{si T por defecto, incluye las etiquetas de países y ciudad en tierra, no funciona en Porcupine}

\item{bw}{Si T gráfico en blanco y negro por default, si F gráfico en color}
}
\description{
Grafica todas las dragas o CTDs de una campaña siempre que se hayan marcado con el PescaWin durante la misma
Requiere que los ficheros de marcas del PescaWin estén en los directorios "C:/GPS/mrk/porcupine/" para Porcupine o "C:/GPS/mrk/Norte/" para Demersales y "C:/GPS/mrk/arsa/" para el Golfo de Cádiz
}
\examples{
\dontrun{
GrafMarksGPS64("N14","Cant")
}
}
\seealso{
Other mapas: 
\code{\link{MapEcol64.camp}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{VedaCigala}},
\code{\link{VedaCigala.out}},
\code{\link{VedasBesugo}},
\code{\link{arMapLansGPS64}()},
\code{\link{armap.camp64}()},
\code{\link{maparea64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()},
\code{\link{maphistal64}()}

Other PescaWin: 
\code{\link{MapLansGPS64}()},
\code{\link{arMapLansGPS64}()}
}
\concept{PescaWin}
\concept{mapas}
