% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HdectoH.R
\name{HdectoH}
\alias{HdectoH}
\title{Transforma hora decimal en formato HH.NN a hora hexadecimal HH.MM
Transforma horas decimales a horas hexadecimales en horas y minutos}
\usage{
HdectoH(x)
}
\arguments{
\item{x}{horas en formato decimal}
}
\description{
Transforma hora decimal en formato HH.NN a hora hexadecimal HH.MM
Transforma horas decimales a horas hexadecimales en horas y minutos
}
\examples{
HdectoH("10.9")
}
\seealso{
Other Conversion unidades: 
\code{\link{gradec}()},
\code{\link{gradms}()}
}
\concept{Conversion unidades}
