% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HextoDec.R
\name{HextoDec}
\alias{HextoDec}
\title{Transforma hora hexadecimal en formato HH.MM a hora decimal H.
Transforma horas y minutos a horas decimales, avisa si el dato es incorrecto, es decir más de 60 minutos}
\usage{
HextoDec(x)
}
\arguments{
\item{x}{horas y minutos en formato HH.MM, vector de caracter}
}
\description{
Transforma hora hexadecimal en formato HH.MM a hora decimal H.
Transforma horas y minutos a horas decimales, avisa si el dato es incorrecto, es decir más de 60 minutos
}
\examples{
HextoDec("10.54")
}
