% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histboxplot64.comp.R
\name{histboxplot64.comp}
\alias{histboxplot64.comp}
\title{Gráficos de boxplot combinados biomasa y número para la serie histórica}
\usage{
histboxplot64.comp(
  gr,
  esp,
  camps,
  zona = "cant",
  dns = "local",
  cor.time = TRUE,
  es = TRUE,
  bw = TRUE,
  ti = TRUE,
  sub = NULL,
  idi = "l",
  ceros = TRUE,
  cex.leg = 1,
  years = TRUE,
  profrange = NA,
  longrange = NA,
  latrange = NA
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 desechos y otros, 9 escoge todos los orgánicos pero excluye desechos}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{Campaña a representar: Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa", Mediterráneo "medi"}

\item{dns}{Elige de dónde se toman los datos "local" del mismo ordenador, "serv" del servidor}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{es}{Si T rótulos gráfico en español, si F en inglés}

\item{bw}{Gráfico en blanco y negro si T o en color si F}

\item{ti}{Añade el nombre de la especie en latín si T, si F no añade título}

\item{sub}{Añade un subtítulo debajo del gráfico, sin texto por defecto}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{ceros}{Por defecto incluye los valores de 0, si F los quita}

\item{cex.leg}{Varía el tamaño de letra de los ejes}

\item{years}{Si T saca los años como nombre de campaña en el eje x}

\item{profrange}{Si c(profmin,profmax) filtra por ese rango de profundidad}

\item{longrange}{Si c(longmin,longmax) filtra por ese rango de longitudes}

\item{latrange}{Si c(latmin,latmax) filtra por ese rango de latitudes}
}
\value{
Crea una gráfica doble de boxplots en biomasa y número
}
\description{
Crea box plots combinados (biomasa + número) con la distribución de la abundancia
para distintas zonas: Porcupine (zona="porc"), el Cantábrico (zona="cant"),
Cádiz (zona="arsa"), y el Mediterráneo (zona="medi").
}
\examples{
histboxplot64.comp(1,50,Nsh[7:27],"cant",years=TRUE)
}
\seealso{
\link{histboxplot64}

Other abunds: 
\code{\link{histboxplot64}()}
}
\concept{abunds}
