% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histboxplot64.R
\name{histboxplot64}
\alias{histboxplot64}
\title{Gráficos de boxplot para la serie histórica incluyendo lances especiales o no y con rangos de profundidad}
\usage{
histboxplot64(
  gr,
  esp,
  camps,
  zona = "porc",
  dns = "local",
  cor.time = TRUE,
  incl2 = TRUE,
  es = T,
  bw = TRUE,
  ti = TRUE,
  sub = NULL,
  out.dat = FALSE,
  ind = "p",
  idi = "l",
  ceros = TRUE,
  cex.leg = 1.1,
  years = TRUE,
  profrange = NA,
  longrange = NA,
  latrange = NA,
  ranglabel = TRUE,
  nlans = TRUE,
  lan.cex = 0.6,
  graf = FALSE,
  psize = 15
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 desechos y otros, 9 escoge todos los orgánicos pero excluye desechos}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{Campaña a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cadiz "arsa"}

\item{dns}{Elige el origen de los datos, si del ordenador ("local") o del servidor ("serv")}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{incl2}{Si F no tiene en cuenta los lances especiales, si T si los tiene en cuenta, pero da problemas por que no puede calcular las abundancias estratificadas}

\item{es}{Si T rotulos gráfico en español, si F en inglés}

\item{bw}{Gráfico en blanco en negro si T o en color si F}

\item{ti}{Añade el nombre de la especie en latín sin T, si F no añade titulo}

\item{sub}{Añade un subtítulo debajo del gráfico, sin texto por defecto.}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{ind}{Parámetro a representar saca los datos en "p"eso o "n"úmero}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{ceros}{por defecto incluye los valores de 0 al calcular los rangos y medianas, si T los quita, Es importante avisarlo en el la explicación de la gráfica}

\item{cex.leg}{Varía el tamaño de letra de los ejes y del número de la leyenda}

\item{years}{Si T saca los años como nombre de campaña en los paneles lattice de campañas}

\item{profrange}{Si c(profmin,profmax) filtra por ese rango de profundidad por defecto NA no filtra profundidades}

\item{longrange}{Si c(longmin,longmax) filtra por ese rango de longitudes por defecto NA no filtra longitudes}

\item{latrange}{Si c(latmin,latmax) filtra por ese rango de latitudes por defecto NA no filtra latitudes}

\item{ranglabel}{Si T incluye texto con los rangos elegidos, is F los omite}

\item{nlans}{T por defecto presenta el número de lances en la campaña por encima del eje x}

\item{lan.cex}{tamaño de las etiquetas del numero de lances por campaña}

\item{graf}{Si F no saca nada, si pones el nombre de un gráfico lo saca saca como archivo png y al final del proceso dice dónde está el mapa con ese nombre:}

\item{psize}{Tamaño del punto del archivo png, en este gráfico por defecto 15}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{escmult}{Varía la relación de tamaño de los puntos con la leyenda y el máximo en los datos}
}
\value{
Si out.dat=TRUE devuelve un data.frame con columnas: lan,lat,long,prof,peso.gr,numero (de individuos entre tmin y tmax),camp, si out.dat=F saca el gráfico en pantalla o como objeto para combinar con otros gráficos con print.trellis
}
\description{
Crea box plots con la distribución de la abundancia, en biomasa o número, no teniendo
en cuenta la ponderación al área (\emph{see details}) para distintas zonas: Porcupine
(zona="porc"), el Cantábrico (zona=cant), Cádiz= (zona=arsa), y el Mediterráneo
(zona=medi).
}
\details{
Estos boxplots pueden ser engañosos y hay que explicar bien en la gráfica sus resultados. Sí se decide eliminar los ceros (=F) hay que tener cuidado porque los
resultados pueden depender mucho del número de lances sin captura y dar resultados muy engañosos al poder compararse datos con distinto numero de lances. Ver en
examples ambas opciones
}
\examples{
histboxplot(1,50,Nsh[7:27],"Cant",years=TRUE)
histboxplot(1,50,Nsh[7:27],"Cant",years=TRUE,ind="n")
histboxplot(1,50,Nsh[7:27],"Cant",years=TRUE,ind="n",ceros=FALSE)
histboxplot(1,50,Nsh[21:40],"Cant",years=T,ind="p",latrange=c(42,43))
histboxplot(1,50,Nsh,"Cant",latrange = c(42,43),graf="Graf")
}
\seealso{
Other abunds: 
\code{\link{histboxplot64.comp}()}
}
\concept{abunds}
