% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lan0.camp.r
\name{lan0.camp}
\alias{lan0.camp}
\title{Lances nulos en una campaña}
\usage{
lan0.camp(camp, zona = "cant", dns = c("local", "serv"))
}
\arguments{
\item{camp}{Campaña a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant" o Golfo de Cádiz "arsa"}

\item{dns}{elige de dónde se toman los datos, del ordenador personal ("local") o del servidor ("serv") si está configurado}
}
\description{
Busca y devuelve los lances nulos de una campaña
}
\examples{
lan0.camp(camp="P01",zona="porc",serv="local")
}
