% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lan2.camp.r
\name{lan2.camp}
\alias{lan2.camp}
\title{Lances especiales de una campaña}
\usage{
lan2.camp(camp, zona = "cant", dns = c("local", "serv"))
}
\arguments{
\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa"}

\item{dns}{elige el origen de los datos, si del ordenador ("local") o del servidor ("serv") si está habilitado}
}
\description{
Busca y devuelve los lances especiales de una campaña
}
\examples{
lan2.camp(camp="P01",zona="porc",serv="local"))
}
