% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_utils.R
\name{list_shapes}
\alias{list_shapes}
\title{Listar shapes disponibles en CampR64}
\usage{
list_shapes()
}
\value{
Data.frame con columnas type y file
}
\description{
Listar shapes disponibles en CampR64
}
