% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListFauna.lan64.r
\name{ListFauna.lan64}
\alias{ListFauna.lan64}
\title{Capturas de un lance específico}
\usage{
ListFauna.lan64(camp, zona = "cant", dns = c("local", "serv"), lan, gr = NA)
}
\arguments{
\item{camp}{Campaña de la que se extrae los datos del lance (lan) un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa"}

\item{dns}{Elige de dónde se toman los datos si del ordenador ("local") o del servidor ("serv")}

\item{lan}{Lance del que se extrae la faunística}

\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados. Si NA, muestra todos los grupos a excepción de 6 desechos y otros no orgánicos}
}
\value{
Devuelve un data.frame con columnas grupo,esp,especie,peso,numero
}
\description{
Muestra un listado de las especies capturadas en peso y número en un lance concreto. Se pueden seleccionar grupos de especies
}
\examples{
ListFauna.lan64("12C","cant","local",1,gr=c(1,2))
}
\seealso{
{\link{ListFauna.camps}}, {\link{ListFaunaTals.camps}}
}
