% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapArsa64.R
\name{MapArsa64}
\alias{MapArsa64}
\title{Mapa base Golfo de Cádiz (ARSA) con selección por regiones (modelo MapNort64)
Requiere objetos 'Arsa.map' y 'Arsa.str' en memoria (data/ del paquete).}
\usage{
MapArsa64(
  xlims = c(-8.149, -5.52),
  ylims = c(35.95, 37.3337),
  lwdl = 1,
  leg = FALSE,
  cuadr = FALSE,
  cuadrMSFD = FALSE,
  ICESrect = FALSE,
  FU = NA,
  ColFU = "chartreuse",
  dens = NA,
  FUsLab = FALSE,
  ax = TRUE,
  bw = FALSE,
  es = TRUE,
  places = TRUE,
  sea_col = "lightblue1",
  land_col = "#D2B48C",
  strata_cols = c("#CFE8FF", "#9CC9FF", "#6AA9FF", "#2E7DD6", "#0F4C81"),
  wmf_file = NULL
)
}
\description{
Mapa base Golfo de Cádiz (ARSA) con selección por regiones (modelo MapNort64)
Requiere objetos 'Arsa.map' y 'Arsa.str' en memoria (data/ del paquete).
}
