% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapEcol.camp64.R
\name{MapEcol.camp64}
\alias{MapEcol.camp64}
\title{Mapa de índices ecológicos para un grupo en una campaña}
\usage{
MapEcol.camp64(
  gr,
  esp = "999",
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  ind = "n",
  indec = "div",
  plot = TRUE,
  bw = FALSE,
  ti = TRUE,
  idi = "l",
  es = TRUE,
  out.dat = FALSE,
  layout = NA,
  cexleg = 1,
  years = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados. 6 Desechos y otros no se incluye en esta función}

\item{esp}{ha de ser 999 cuando se quiere incluir todas las especies del grupo, o elegir todas las especies deseadas con los codigos de las especies}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Zona: \code{"cant"}, \code{"porc"}, \code{"arsa"}, \code{"medi"}.}

\item{dns}{Origen de datos: \code{"local"} o \code{"serv"}.}

\item{ind}{Elige el valor (n)úmero o (p)eso sobre el que se calculan los índices de diversidad, dominancia....}

\item{indec}{Elige el índice ecológico a representar: opciones disponibles: Shannon-Wiener: 'div', Número de especies: 'nesp' y Diversidad Simpson: 'simp'.}

\item{plot}{Si T saca un gráfico en pantalla}

\item{ti}{Añade el nombre de la especie en latín sin T, si F no añade titulo}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{es}{Si T saca los titulos y rotulos en español, si F en inglés}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}

\item{cexleg}{Varía el tamaño de letra de los ejes y del número de la leyenda}

\item{years}{Si T saca los años como nombre de campaña en los paneles lattice de campañas}
}
\value{
Saca el mapa de diversidad en la campaña seleccionada.
}
\description{
Utiliza los datos del Camp representar la variación geográfica de los índices ecológicos
}
\examples{
dumbecol<-MapEcol.camp(1,999,Nsh[25:30],"Cant",ind="n",bw=TRUE,indec="simp",out.dat=TRUE,layout=c(2,3))
dumbecol$estrato<-cut(dumbecol$prof,c(0,70,120,200,500,900),c("A1","A","B","C","D"))
lattice::bwplot(numbesp~estrato|camp,dumbecol,horizontal=FALSE,main="Número de especies",xlab="Estrato")
lattice::bwplot(div~estrato|camp,dumbecol,horizontal=FALSE,main="Shannon Wiener",xlab="Estrato")
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{armap.tot64}()},
\code{\link{maparea64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()},
\code{\link{maphistal64}()}

Other ecologia: 
\code{\link{ecolgr.camp64}()},
\code{\link{ecolmatrix.camp64}()}
}
\concept{ecologia}
\concept{mapas}
