% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapLansGPS64.r
\name{MapLansGPS64}
\alias{MapLansGPS64}
\title{Gráfico con los lances en segmentos}
\usage{
MapLansGPS64(
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  Nlans = FALSE,
  incl0 = FALSE,
  xlims = NA,
  ylims = NA,
  places = TRUE,
  es = T,
  bw = FALSE,
  ax = T
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa"}

\item{dns}{Elige si tonma los datos del ordenador "local" o del servidor "serv"}

\item{Nlans}{si T pone el número de lance junto al segmento}

\item{incl0}{Si T se incluyen los lances nulos}

\item{xlims}{Define los limites longitudinales del mapa, si se deja en NA toma los límites de long del área definida en la campaña}

\item{ylims}{Define los limites latitudinales del mapa, si se deja en NA toma los límites de lat del área definida en la campaña}

\item{places}{si T por defecto, incluye las etiquetas de países y ciudad en tierra, no funciona en Porcupine}

\item{bw}{Si T gráfico en blanco y negro por default, si F gráfico en color}

\item{ax}{Si T saca los ejes x e y}
}
\value{
Devuelve un data.frame con datos de cada lance, las variables dependen de la selección de hidro y redux
}
\description{
gráfica todos los lances de una campaña con las distancias recorridas en cada uno de ellos
}
\examples{
\dontrun{
MapLansGPS64("12C","cant")
}
}
\seealso{
\link{datlan.camp64}, \link{qcdistlan.camp64}

Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol64.camp}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{VedaCigala}},
\code{\link{VedaCigala.out}},
\code{\link{VedasBesugo}},
\code{\link{arMapLansGPS64}()},
\code{\link{armap.camp64}()},
\code{\link{maparea64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()},
\code{\link{maphistal64}()}

Other PescaWin: 
\code{\link{GrafMarksGPS}()},
\code{\link{arMapLansGPS64}()}
}
\concept{PescaWin}
\concept{mapas}
