% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapMedit.R
\name{MapMedit}
\alias{MapMedit}
\title{Mapa del Mediterráneo Ibérico completo}
\usage{
MapMedit(
  xlims = c(-5.7, 5),
  ylims = c(35, 43),
  lwdl = 1,
  cuadr = FALSE,
  cuadrMSFD = FALSE,
  ICESrect = FALSE,
  ICESrectcol = gray(0.2),
  ICESlab = FALSE,
  ICESlabcex = 0.5,
  bw = F,
  ax = TRUE,
  wmf = FALSE,
  es = TRUE,
  places = TRUE
)
}
\arguments{
\item{xlims}{Define los limites longitudinales del mapa, los valores por defecto son los del total del área de la campaña}

\item{ylims}{Define los limites latitudinales del mapa, los valores por defecto son los del total del área de la campaña}

\item{lwdl}{Ancho de las líneas del mapa}

\item{cuadr}{Si T saca las cuadrículas de 5x5 millas naúticas}

\item{cuadrMSFD}{Si T dibuja caudrícula de 10 millas naúticas utilizada para la evaluación de la estrategia marina (MSFD)}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{ICESrectcol}{Color para los rectángulos ICES}

\item{ICESlab}{Si T incluye las etiquetas de los rectángulos ICES}

\item{ICESlabcex}{tamaño del ICESlab en cex, .5 por defecto subirlo si se quiere más grande}

\item{bw}{si T mapa con tierra en gris, si F tierra en color}

\item{ax}{Si T saca los ejes x e y}

\item{wmf}{Si T saca a fichero metafile Meditconc.emf}

\item{es}{Si T saca titulos y ejes en español}

\item{places}{Si T saca ciudades y puntos geográficos de referencia}
}
\value{
Saca en pantalla el mapa y es utilizada por otras funciones
}
\description{
Función auxiliar para sacar mapas de la campaña MEDITS
}
\examples{
MapMedit()
}
\concept{Medits}
\concept{mapas base}
