% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapNort64.R
\name{MapNort64}
\alias{MapNort64}
\title{Mapa base Mar Cantábrico y costa norte de España}
\usage{
MapNort64(
  lwdl = 0.5,
  cuadr = FALSE,
  cuadrcol = gray(0.4),
  cuadrMSFD = FALSE,
  ICESrect = FALSE,
  ICESrectcol = gray(0.2),
  ICESlab = FALSE,
  Sects = TRUE,
  SecLty = 2,
  ICESlabcex = 0.7,
  leg = FALSE,
  bw = FALSE,
  es = FALSE,
  ax = TRUE,
  strat = FALSE,
  places = FALSE,
  FU = NA,
  ColFU = "chartreuse",
  FUsLab = FALSE,
  dens = 20,
  xlims = c(-9.9, -1.4),
  ylims = c(41.91, 44.19)
)
}
\arguments{
\item{lwdl}{grosor de línea de costa}

\item{cuadr}{cuadrícula de muestreo}

\item{cuadrcol}{color de la cuadrícula}

\item{cuadrMSFD}{cuadrícula MSFD}

\item{ICESrect}{rectángulos ICES}

\item{ICESrectcol}{color rectángulos ICES}

\item{ICESlab}{etiquetas rectángulos ICES}

\item{Sects}{dibujar sectores de la campaña}

\item{SecLty}{tipo de línea de los sectores}

\item{ICESlabcex}{tamaño etiquetas ICES}

\item{leg}{leyenda de estratos de profundidad}

\item{bw}{blanco y negro}

\item{es}{etiquetas en español}

\item{ax}{dibujar ejes}

\item{strat}{sombrear estratos de profundidad}

\item{places}{dibujar ciudades}

\item{FU}{Functional Units a dibujar}

\item{ColFU}{color de las FU}

\item{FUsLab}{etiquetas de las FU}

\item{dens}{densidad de tramado de las FU}

\item{xlims}{límites de longitud}

\item{ylims}{límites de latitud}
}
\description{
Dibuja el mapa base del área del Cantábrico usando sf para los polígonos
y maps::map para gestionar automáticamente el aspecto geográfico correcto.
}
