% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPorc64.R
\name{MapPorc64}
\alias{MapPorc64}
\title{Mapa base Banco de Porcupine}
\usage{
MapPorc64(
  xlims = c(-15.5, -10.5),
  ylims = c(50.5, 54.5),
  lwdl = 1,
  cuadr = FALSE,
  ICESrect = FALSE,
  label = FALSE,
  ax = TRUE,
  bw = FALSE,
  es = FALSE,
  places = FALSE,
  corners = FALSE
)
}
\description{
Mapa base Banco de Porcupine
}
