% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsorteo.r
\name{mapsorteo}
\alias{mapsorteo}
\title{Mapa de las cuadrículas sorteadas para la campaña Porcupine

Mapa del resultado de un sorteo de cuadriculas de Porcupine, con simbolos diferentes para cada sector/estrato}
\usage{
mapsorteo(
  x,
  wmf = FALSE,
  ax = TRUE,
  nomfic = "porco.emf",
  lwdl = 0.5,
  colo = NA,
  add = FALSE,
  labs = FALSE,
  col = 2,
  cex = 0.9
)
}
\arguments{
\item{x}{El resultado de un sorteo mediante la función sorteo(). Puede ser un objeto con la misma estructura que el resultado de utilizar sorteo {\link{sorteo}}}

\item{wmf}{si T guarda el resultado en un metafile, si F saca el gráfico en pantalla}

\item{ax}{si T saca los ejes en el gráfico, si F queda solo el grid, útil para sacar metafiles a utilizar en el PescaWin}

\item{nomfic}{Especifica el nombre del fichero metafile de salida por defecto Porco.emf si wmf es T}

\item{lwdl}{Ancho de las líneas que marcan los estratos en el mapa (2 suelen verse mejor)}

\item{colo}{Color de las líneas del mapa}

\item{add}{Si T añade los puntos al gráfico en pantalla, si F dibuja un mapa nuevo}

\item{labs}{Si F dibuja puntos, si T saca el número de las cuadriculas en su cuadrícula correspondiente}

\item{col}{Color de los puntos del sorteo en x}

\item{cex}{Tamaño del texto de las etiquetas de cuadrícula si labs=TRUE}
}
\description{
Mapa de las cuadrículas sorteadas para la campaña Porcupine

Mapa del resultado de un sorteo de cuadriculas de Porcupine, con simbolos diferentes para cada sector/estrato
}
\seealso{
{\link{sorteo}}
}
