% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsorteo64.R
\name{mapsorteo64}
\alias{mapsorteo64}
\title{Mapa de las cuadrículas sorteadas para la campaña Porcupine}
\usage{
mapsorteo64(
  x,
  wmf = FALSE,
  ax = TRUE,
  nomfic = "porco.emf",
  lwdl = 0.5,
  colo = NA,
  add = FALSE,
  labs = FALSE,
  col = 2,
  cex = 0.9
)
}
\arguments{
\item{x}{Resultado de sorteo64()}

\item{wmf}{Si TRUE guarda en metafile Windows}

\item{ax}{Si TRUE muestra ejes}

\item{nomfic}{Nombre del fichero metafile (default "porco.emf")}

\item{lwdl}{Ancho de líneas de estratos}

\item{colo}{Color de líneas del mapa}

\item{add}{Si TRUE añade al gráfico existente}

\item{labs}{Si TRUE muestra números de cuadrícula, si FALSE muestra puntos}

\item{col}{Color de los puntos/etiquetas}

\item{cex}{Tamaño de etiquetas si labs=TRUE}
}
\description{
Mapa del resultado de un sorteo de cuadrículas de Porcupine
}
\seealso{
{\link{sorteo64}}
}
