% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchHidro.ctd64.R
\name{MatchHidro.ctd64}
\alias{MatchHidro.ctd64}
\title{Mapa de estaciones hidrológicas (CTDs) en una campaña}
\usage{
MatchHidro.ctd64(
  camp,
  zona = "porc",
  dns = c("local", "serv"),
  plot = TRUE,
  subtit = TRUE,
  bw = FALSE,
  ti = TRUE,
  es = TRUE,
  out.dat = FALSE,
  cex.pt = 1,
  cexleg = 1,
  years = TRUE,
  observ = FALSE,
  estn = FALSE,
  lanctd = FALSE,
  graf = FALSE,
  xpng = 1200,
  ypng = 800,
  ppng = 15,
  cuts = 4,
  restore.par = TRUE
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{dns}{Elige si los datos se toman del ordenador ("local") o del servidor ("serv")}

\item{plot}{Saca el gráfico (si T) o si se salva como objeto se puede componer para componer con otros gráficos de lattice (F)}

\item{ti}{Añade la Campaña, si F no añade titulo}

\item{es}{Si T saca los titulos y rotulos en español, si F en inglés}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{cex.pt}{Varía el tamaño de los puntos en los gráficos}

\item{cexleg}{varía el tamaño del texto de la leyenda y los ejes}

\item{years}{Si T saca los años como nombre de campaña en los paneles lattice de campañas}

\item{graf}{si F no el gráfico va a pantalla, si nombre fichero va a fichero en el directorio en que está wdf}

\item{xpng}{width archivo png si graf es el nombre del fichero}

\item{ypng}{height archivo png si graf es el nombre del fichero}

\item{ppng}{points png archivo si graf es el nombre del fichero}

\item{ind}{Elige el valor (t)emperatura o (s)alinidad}

\item{sub}{Si T añade el parámetro (temperatura, salinidad) como subtítulo bajo el gráfico en inglés o español}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}
}
\value{
Saca el mapa de diversidad en la campaña seleccionada.
}
\description{
Utiliza los datos del Camp representar la variación geográfica de los datos hidrologicos, temperatura y salinidad
}
\examples{
\dontrun{
MatchHidro.ctd64("N24","cant","local",ti=T)
}
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol64.camp}()},
\code{\link{MapLansGPS64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{VedaCigala}},
\code{\link{VedaCigala.out}},
\code{\link{VedasBesugo}},
\code{\link{arMapLansGPS64}()},
\code{\link{armap.camp64}()},
\code{\link{maparea64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()},
\code{\link{maphistal64}()}
}
\concept{hidrología}
\concept{mapas}
