% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msh.R
\docType{data}
\name{Msh}
\alias{Msh}
\title{Serie histórica MEDITS (Msh)}
\format{
An object of class \code{character} of length 18.
}
\usage{
Msh
}
\description{
Vector con los códigos de campaña MEDITS desde 1994 (M94) hasta 2011 (M11).
}
\keyword{datasets}
