% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NepFU30.camp64.R
\name{NepFU30.camp64}
\alias{NepFU30.camp64}
\title{Información anual sobre cigala para la FU 30 en el Golfo de Cádiz}
\usage{
NepFU30.camp64(
  camp,
  zona = "arsa",
  dns = c("local", "serv"),
  trimes = 4,
  plot = TRUE,
  es = FALSE,
  ti = TRUE,
  ICESlab = FALSE,
  ceros = T,
  leg = T,
  escmult = 0.25,
  cexleg = 1,
  ICESrectcol = 1,
  ICESrect = TRUE,
  places = TRUE,
  FU = "FU30",
  ColFU = "white",
  dens = 20,
  out.dat = TRUE,
  bw = FALSE,
  graf = FALSE,
  xpng = 1100,
  ypng = 1100,
  ppng = 20
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Arsa "1XX"}

\item{zona}{Sólo disponible para el Golfo de Cádiz "arsa"}

\item{dns}{elige datos de ordenador "local" o del servidor "serv"}

\item{plot}{Saca el gráfico (T) o lo omite para dejar sólo los datos (F)}

\item{es}{si T letreros en español, si F en inglés (F por defecto)}

\item{ICESlab}{Si T incluye las etiquetas de los rectángulos ICES}

\item{ICESrectcol}{Color para los rectángulos ICES}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{places}{Si T saca etiquetas de principales ciudades en el mapa, si F se omiten los letreros}

\item{FU}{pinta los límites de la FU30 en color rojo}

\item{ColFU}{color de relleno de la FU, blanco por defecto, con dens=0 se omite el relleno.}

\item{dens}{valores de 20 hace que el color permita ver el fondo con los ICES rects y la batimetría, más lo deja en color sólido, 0 quita el relleno del todo.}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos de abundancia de cigala en cada FU con datos de biomasa, numero, desviaciones estándar y número de lances en cada FU.}

\item{graf}{Si FALSE el gráfico va a pantalla; si es un string guarda como PNG con ese nombre}

\item{xpng}{Anchura del PNG en píxeles (default 1100)}

\item{ypng}{Altura del PNG en píxeles (default 1100)}

\item{ppng}{Pointsize del PNG (default 20)}

\item{year}{si T incluye una columna con el año al final de los datos}

\item{plotnep}{si T presenta todos los lances en los que ha habido cigala en la campaña, además de los lances en cada FU, con o sín captura. Si F sólo saca los lances en cada FU sin marcar si ha habido cigala o no.}

\item{ICESlabcex}{tamaño del ICESlab en cex, .5 por defecto subirlo si se quiere más grande}
}
\value{
Produce un gráfico con los lances en los que ha habido cigala en el lance y especialmente los lances en cada FU dentro de Demersales FU25,FU30 y FU31
}
\description{
Función de acceso a datos:
Extrae las características de los lances para una campaña determinada y las capturas de cigala por unidad funcional
}
\details{
Un problema que ocurre al utilizar el CampR con ficheros dbf de las primeras campañas
puede ser que al fichero lanceXXX.dbf le falte algún campo, habitualmente
el campo \strong{ESTN} utilizado en las últimas versiones del \strong{CAMP} para ligar lances con las estaciones de CTD.
El error usual es \strong{$ operator is invalid for atomic vectors}
Si se detecta este error revisar la estructura de lanceXXX.dbf con la de
otros ficheros de lances de los últimos años
}
\examples{
\dontrun{
  NepFU30.camp64(camp="112",zona="arsa")
  NepFU30.camp64(camp="124",zona="arsa",dns="local",graf="FU30_2024")
  }
}
\seealso{
Other mapas, NEP: 
\code{\link{NepFU25.camp64}()},
\code{\link{NepFU26.camp64}()},
\code{\link{NepFU31.camp64}()}
}
\concept{mapas, NEP}
