% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nsh.R
\docType{data}
\name{Nsh}
\alias{Nsh}
\title{Serie histórica Demersales Norte (Nsh)}
\format{
An object of class \code{character} of length 29.
}
\usage{
Nsh
}
\description{
Vector con los códigos de campaña del cuarto trimestre en Cantábrico y Galicia.
}
\concept{Series históricas}
\keyword{datasets}
