% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psh.R
\docType{data}
\name{Psh}
\alias{Psh}
\title{Serie histórica Porcupine (Psh)}
\format{
An object of class \code{character} of length 25.
}
\usage{
Psh
}
\description{
Vector con los códigos de campaña realizados en Porcupine.
}
\seealso{
Other series históricas: 
\code{\link{As1}},
\code{\link{As2}}
}
\concept{series históricas}
\keyword{datasets}
