% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcdtallrev64.camp.R
\name{qcdtallrev64.camp}
\alias{qcdtallrev64.camp}
\title{Revisión de los histogramas de tallas de todas las especies de una campaña}
\usage{
qcdtallrev64.camp(
  camp = "N12",
  zona = "cant",
  dns = c("local", "serv"),
  nlans = 2
)
}
\arguments{
\item{camp}{Campaña: Demersales "NXX", Porcupine "PXX", Arsa "1XX"/"2XX"}

\item{zona}{Zona: "cant", "porc", "arsa"}

\item{dns}{Origen de datos: "local" o "serv"}

\item{nlans}{Mínimo de lances en que ha aparecido la especie para mostrarla (default 2)}
}
\description{
Revisa la distribución de tallas de todas las especies de peces que hayan aparecido
en una campaña, una por una en pantalla. Se puede salir cerrando el gráfico.
}
\examples{
\dontrun{
qcdtallrev64.camp(camp="N25", zona="cant", dns="local", nlans=2)
}
}
\seealso{
\code{\link{qcdtall.camp64}}

Other Control de calidad: 
\code{\link{qcLW64.camp}()},
\code{\link{qcLWbucl64.camp}()},
\code{\link{qcTalPez64.camp}()},
\code{\link{qcTalPez64.lan}()},
\code{\link{qcdistlan.camp64}()},
\code{\link{qcdtall64.camp}()},
\code{\link{qclandatr.camp64}()}
}
\concept{Control de calidad}
