% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcLanDatr.camp64.R
\name{qclandatr.camp64}
\alias{qclandatr.camp64}
\title{Comprueba la distancia recorrida en los lances y la consistencia con recorrido y la velocidad en los datos del CAMP}
\usage{
qclandatr.camp64(camp, zona = "cant", dns = "local")
}
\arguments{
\item{camp}{campaña a revisar los datos en formato Camp Xyy}

\item{zona}{Origen de bases de datos: "cant" cantábrico, "porc" Porcupine, "arsa" para el Golfo de Cádiz y "medit" para MEDITS}

\item{plots}{si T abre una pantalla con cuatro gráficos de comparacion entre camp y DATRAS}
}
\value{
Devuelve un gráfico con cuatro plots puertas~prof, calones~prof, vertical~prof y lances en área
}
\description{
Sirve para control de calidad y asegurarse que los datos de distancias, posiciones, y las horas de largada y virada son correctos y no salen del estándar de largar y virar de día
}
\examples{
\dontrun{
qclandatr.camp64("N23","cant","local")
qclandatr.camp64("216","arsa","local")
}
}
\seealso{
\link{MapLansGPS64}

Other Control de calidad: 
\code{\link{qcLW64.camp}()},
\code{\link{qcLWbucl64.camp}()},
\code{\link{qcTalPez64.camp}()},
\code{\link{qcTalPez64.lan}()},
\code{\link{qcdistlan.camp64}()},
\code{\link{qcdtall64.camp}()},
\code{\link{qcdtallrev64.camp}()}
}
\concept{Control de calidad}
