% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_utils.R
\name{read_shape}
\alias{read_shape}
\title{Leer un shape interno de CampR64 como objeto sf}
\usage{
read_shape(type, file, quiet = TRUE)
}
\arguments{
\item{type}{Tipo: "ices", "cant", "porc", "arsa", "iberia", "bathy"}

\item{file}{Nombre del fichero (con extension .gpkg)}

\item{quiet}{Suprimir mensajes (default TRUE)}
}
\value{
Objeto sf
}
\description{
Leer un shape interno de CampR64 como objeto sf
}
