% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCampDBF.R
\name{readCampDBF}
\alias{readCampDBF}
\title{Leer un fichero DBF de CampR64 según zona y origen}
\usage{
readCampDBF(
  tipo,
  zona = c("cant", "porc", "arsa", "medi"),
  camp = NULL,
  dns = c("local", "serv")
)
}
\description{
Leer un fichero DBF de CampR64 según zona y origen
}
