% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacagrid64.R
\name{sacagrid64}
\alias{sacagrid64}
\title{Saca el grid de cuadrículas para Porcupine}
\usage{
sacagrid64(color = 2, cexlab = 0.7, etiq = TRUE)
}
\arguments{
\item{color}{Color de las etiquetas (default 2)}

\item{cexlab}{Tamaño de las etiquetas (default 0.7)}

\item{etiq}{Si TRUE muestra etiquetas de número de cuadrícula}
}
\value{
Data.frame con columnas x, y, pt, strat
}
\description{
Crea el grid de cuadrículas de Porcupine con los puntos centrales de cada
una de las cuadrículas (5x5 millas náuticas) disponibles en la campaña
}
\seealso{
{\link{sorteo64}}, {\link{mapsorteo64}}
}
