% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacagridNort64.R
\name{sacagridNort64}
\alias{sacagridNort64}
\title{Saca el grid de cuadrículas para Demersales Norte}
\usage{
sacagridNort64(color = 2, cexlab = 0.7, etiq = TRUE)
}
\arguments{
\item{color}{Color de las etiquetas (default 2)}

\item{cexlab}{Tamaño de las etiquetas (default 0.7)}

\item{etiq}{Si TRUE muestra etiquetas de número de cuadrícula}
}
\value{
Data.frame con columnas x, y, pt, strat
}
\description{
Crea el grid de cuadrículas del Cantábrico/Galicia (5x5 millas náuticas)
}
\seealso{
{\link{sacagrid64}}
}
