% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sefos64.camp.R
\name{sefos64.camp}
\alias{sefos64.camp}
\title{Crea datos en formato SEFOS a partir de Camp, dns}
\usage{
sefos64.camp(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  plus = 8,
  cor.time = TRUE,
  AltAlk = NA,
  incl2 = FALSE,
  mediahora = 2
)
}
\arguments{
\item{gr}{Grupo de organismos, 1 peces 2 crustaceos, 3 moluscos, 4 equinodermos, 5 inv..}

\item{esp}{Código de especie camp}

\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "cant", Golfo de Cádiz "arsa" (únicamente para sacar datos al IBTS, no gráficos)gr Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{dns}{Elige de dónde se toman los datos, si del ordenador "local" o del servidor "serv"}

\item{plus}{Cuál es la edad plus (en el formato SEFOS las opciones son 3+, 5+ o 12+ según el ciclo vital de la especie.}

\item{cor.time}{por defecto T, corrección del tiempo de arrastre al calcular las abundancias}

\item{AltAlk}{ALK alternativa tomada de un fichero de edad del camp edadxyy.dbf sin ruta ni extensión}

\item{incl2}{Si T se incluyen los lances extra no incluidos para las abundancias o biomasas estratificadas}

\item{mediahora}{=2, las abundancias tienen que ser por hora, commo el lance dura media hora se pone a dos, en el caso de ARSA debería ser 1}
}
\value{
A data.table con el formato de datos de SEFOS para la evaluación de caballa
}
\description{
Function to complete HH with ICESrect and area
}
\examples{
sefos64.camp(1,74,"N23","cant","local",plus=3)
}
