% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_campR64_root.R
\name{set_campR64_root}
\alias{set_campR64_root}
\title{Establecer la ruta raíz de los datos para CampR64}
\usage{
set_campR64_root(ruta)
}
\arguments{
\item{ruta}{Cadena con la ruta raíz donde se encuentran las carpetas
de datos de CampR64.}
}
\value{
Invisiblemente TRUE si la ruta es válida.
}
\description{
Permite cambiar dinámicamente entre diferentes ubicaciones de datos,
como "C:/camp" o "Z:/camp/datos". La función valida que la ruta existe
y que contiene las subcarpetas esperadas (arsa, cant, porc, medits, camp).
}
\examples{
\dontrun{
set_campR64_root("C:/camp")
set_campR64_root("Z:/camp/datos")
}

}
