% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorteo64.R
\name{sorteo64}
\alias{sorteo64}
\title{Sorteo de distribución de lances en Porcupine}
\usage{
sorteo64(buffer = 0.1177, srv.dsg = Porc.dsg)
}
\arguments{
\item{buffer}{Distancia mínima entre cuadrículas contiguas}

\item{srv.dsg}{Diseño del muestreo: data.frame con columnas strat y stations}
}
\value{
Data.frame con columnas x, y, pt, strat
}
\description{
Realiza el sorteo de las cuadrículas de muestreo en la campaña de Porcupine
con el número correspondiente de estaciones por sector y estrato y aplicando
una distancia mínima entre lances contiguos
}
\seealso{
{\link{mapsorteo64}}, {\link{sacagrid64}}
}
