% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strmean.camps64.r
\name{strmean.camps64}
\alias{strmean.camps64}
\title{Media estratificada total del área de la especie X en varias campañas}
\usage{
strmean.camps64(x, sector, area, w = rep(1, length(x)), camps)
}
\arguments{
\item{x}{Es el vector de abundancias en peso o número de más de una campaña}

\item{sector}{Es la asignación al sector de cada muestra igual que en strmean {\link{strmean}} pero ha de tener dos caracteres, primero el sector geográfico y segundo el estrato batimétrico}

\item{area}{El área del sector al que pertenece cada lance}

\item{w}{Es un factor de ponderación para cada uno de los valores de x. Sirve para la función de la librería boot boot::boot(x,strmean,R,stype="f",Porc.map=sector,sector=sector,area=area,camps=camps)}

\item{camps}{Son las campañas a que pertenece cada lance}
}
\value{
Devuelve la media estratificadda ponderada al área para cada una de las campañas solicitadas
}
\description{
Similar a strmean.dtt media ponderada al tamaño del estrato pero para varias campañas
}
\seealso{
Other Calculos internos no mostrado: 
\code{\link{strmean.dtt64}()},
\code{\link{strmean64}()}
}
\concept{Calculos internos no mostrado}
