% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strmean.dtt64.r
\name{strmean.dtt64}
\alias{strmean.dtt64}
\title{Medias intraestrato y estratificadas por estratos geográfico y batimétrico}
\usage{
strmean.dtt64(x, sector, area, w = rep(1, length(x)))
}
\arguments{
\item{x}{Es el vector de abundancias en peso o número}

\item{sector}{Es la asignación al sector de cada muestra igual que en strmean {\link{strmean}} pero ha de tener dos caracteres, primero el sector geográfico y segundo el estrato batimétrico}

\item{area}{El área del sector}

\item{w}{Es un factor de ponderación para cada uno de los valores de x. Sirve para la función de la librería boot boot::boot(x,strmean,R,stype="f",Porc.map=sector,sector=sector,area=area)}
}
\value{
Devuelve la media estratificadda ponderada al área dentro de cada estrato y subestrato y la total del área
}
\description{
Función interna para cálculos y aplicación de bootstrap
}
\seealso{
Other Calculos internos no mostrado: 
\code{\link{strmean.camps64}()},
\code{\link{strmean64}()}
}
\concept{Calculos internos no mostrado}
