% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strmean64.R
\name{strmean64}
\alias{strmean64}
\title{Media estratificada total del area de una especie}
\usage{
strmean64(x, sector, area, w = rep(1, length(x)))
}
\arguments{
\item{x}{Es el vector de abundancias en peso o número}

\item{sector}{Es la asignación al sector de cada muestra}

\item{area}{El área del sector}

\item{w}{Es un factor de ponderación para cada uno de los valores de x. Sirve para la función de la librería boot boot::boot(x,strmean,R,stype="f",Porc.map=sector,sector=sector,area=area)}
}
\value{
Devuelve la media estratificadda ponderada al área de los estratos
}
\description{
Función interna para cálculos y aplicación de bootstrap
}
\seealso{
Other Calculos internos no mostrado: 
\code{\link{strmean.camps64}()},
\code{\link{strmean.dtt64}()}
}
\concept{Calculos internos no mostrado}
