% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talbox64.camp.R
\name{talbox64.camps}
\alias{talbox64.camps}
\title{Abundancia estratificada para un rango de talla}
\usage{
talbox64.camps(
  gr,
  esp,
  camps,
  zona = "cant",
  dns = c("local", "Serv"),
  notch = TRUE,
  outline = FALSE,
  varwidth = T,
  cor.time = TRUE,
  boxplot = T,
  excl.sect = NA,
  years = TRUE,
  mult = 100,
  ti = TRUE,
  las = 2,
  es = FALSE,
  bw = TRUE,
  idi = "l",
  cexleg = 1
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{campañas (años) a representar en el mapa: Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "cant", Golfo de Cádiz "arsa" (únicamente para sacar datos al IBTS, no gráficos)gr Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{dns}{Elige de dónde se toman los datos, si del ordenador "local" o del servidor "serv"}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{years}{Si T muestra los datos por años, si F por campañas (siguiendo el formato del parámetro camps)}

\item{mult}{por defecto 100 para evitar que la distribución en números decimales de número haga que no salgan datos}

\item{ti}{Si T en el gráfico muestra el nombre de la especie y el rango de tallas comprendido}

\item{las}{Controla el sentido de las etiquetas del gráfico, 2 perpendicular al eje, mejor para etiquetas de años}

\item{es}{Si T gráfico en castellano, si F gráfico en inglés}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}
}
\value{
Devuelve un vector con nombre con el número estratificado del rango de tallas deseados por campaña/año. Si se solicita plot=TRUE saca un gráfico de barras que muestra la abundancia por año. En peso sólo saca los resultados para una especie.
}
\description{
Extrae los datos de abundancia de una especie o conjunto de especies con un rango de tallas determinado a partir de las distribuciones de talla.También puede mostrar los datos de biomasa a partir de la relación talla-peso
}
\examples{
talbox.camps(2,19,Psh,"Porc",varwidth=T,notch=T)
}
\seealso{
{\link{dattal.camp64}}
}
