% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talpes.camp64.R
\name{talpes.camp64}
\alias{talpes.camp64}
\title{Parámetros a y b de la relación talla-peso (CampR64)}
\usage{
talpes.camp64(grupo, especie, zona = "cant", dns = c("local", "serv"))
}
\arguments{
\item{grupo}{Integer (1 peces, 2 crustáceos, 3 moluscos, 4 equinodermos, 5 invertebrados).}

\item{especie}{Integer o string; código (se convierte a integer).}

\item{zona}{"cant","porc","arsa","medi". Por defecto "cant".}

\item{dns}{"local" (default) o "serv".}
}
\value{
numeric length 2: c(a, b)
}
\description{
Parámetros a y b de la relación talla-peso (CampR64)
}
