% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unid.camp64.R
\name{unid.camp64}
\alias{unid.camp64}
\title{Unidades en que se mide una especie}
\usage{
unid.camp64(gr, esp, zona = "cant", dns = c("local", "serv"))
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}
}
\description{
Da información de las unidades en que se miden las especies y permite añadir datos a los gráficos
}
