% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unid64.camp.R
\name{unid64.camp}
\alias{unid64.camp}
\title{Unidades en que se mide una especie}
\usage{
unid64.camp(gr, esp, zona = "cant", dns = c("local", "serv"))
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{zona}{Zona de la que se leen las especies. Determina la estructura de la tabla de especies, que difiere entre ARSA y el resto (Porcupine/Cantábrico): "cant", "porc", "arsa", "medi"}

\item{dns}{Origen de las bases de datos: "local" en el ordenador, "serv" en el servidor}
}
\description{
Da información de las unidades en que se miden las especies y permite añadir datos a los gráficos
}
