% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VedaCigala.out.R
\docType{data}
\name{VedaCigala.out}
\alias{VedaCigala.out}
\title{Polígono zona ampliada veda cigala Porcupine}
\format{
A data.frame with long and lat points defining the area
}
\source{
Generado con \code{data-raw/VedasCigala.R}
}
\usage{
data(VedaCigala.out)
}
\description{
VedaCigala.out: Zona ampliada de la zona de veda de Cigala en Porcupine
}
\details{
Polígono que delimita una zona ampliada más grande que la zona de veda de la cigala en el Porcupine Bank, la zona de veda fue propuesta por los armadores españoles, y puesta en vigor en 2010 desde el 1 de mayo al 31 de julio. Desde entonces ha estado vigente en diversos periodos, pero al menos todo el mes de mayo desde 2013 a 2016.
}
\examples{
MapPorc64(); lines(lat ~ long, VedaCigala, col = 2, lwd = 2)
lines(lat ~ long, VedaCigala.out, col = "blue", lwd = 2)
}
\references{
Medidas de gestión del stock de Cigala del banco de Porcupine FU 16 \url{http://www.nwwac.org/_fileupload/Opinions and Advice/Year 12/Dictamen_CCANOC_Gestion_Cigala_Porcupine_UF16_2017_ES.pdf}
}
\seealso{
\code{\link{VedaCigala}}

Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol64.camp}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{VedaCigala}},
\code{\link{VedasBesugo}},
\code{\link{arMapLansGPS64}()},
\code{\link{armap.camp64}()},
\code{\link{maparea64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()},
\code{\link{maphistal64}()}

Other Porcupine: 
\code{\link{VedaCigala}},
\code{\link{maparea64}()}
}
\concept{Porcupine}
\concept{mapas}
\keyword{datasets}
