% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VedasBesugo.R
\docType{data}
\name{VedasBesugo}
\alias{VedasBesugo}
\title{Polígonos áreas de veda besugo Cantábrico}
\format{
A data.frame with long and lat points defining the closed areas, polygons separated by NA
}
\source{
Generado con \code{data-raw/VedasBesugo.R}
}
\usage{
data(VedasBesugo)
}
\description{
VedasBesugo: Vedas Besugo Cantábrico y Galicia
}
\details{
Polígonos que delimitan las zonas de veda de juveniles de besugo en el Cantábrico, la zona de veda fue propuesta por los armadores españoles en 2019, sin llegar a entrar en vigor. Los cinco polígonos van separados por filas NA, de modo que se pueden dibujar directamente con \code{lines} o \code{polygon}.
}
\examples{
MapNort64(); lines(lat ~ long, VedasBesugo, col = 2, lwd = 2)
}
\references{
Propuesta de vedas protección juveniles besugo \url{https://www.mapa.gob.es/pesca/participacion-publica/detalle/nuevas_vedas_cantabrico}
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol64.camp}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{VedaCigala}},
\code{\link{VedaCigala.out}},
\code{\link{arMapLansGPS64}()},
\code{\link{armap.camp64}()},
\code{\link{maparea64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()},
\code{\link{maphistal64}()}
}
\concept{Cantábrico Galicia}
\concept{mapas}
\keyword{datasets}
