% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arMapLansGPS64.R
\name{arMapLansGPS64}
\alias{arMapLansGPS64}
\title{Combinación en dos mapas de armap.camp con etiquetas de los lance y mapa con los lances en segmentos}
\usage{
arMapLansGPS64(
  camp,
  zona = "porc",
  dns = c("local", "serv"),
  incl0 = FALSE,
  xlims = NA,
  ylims = NA,
  col = 2,
  lwd = 2,
  cexlab = 0.6,
  places = TRUE,
  Nlans = FALSE,
  es = TRUE,
  bw = FALSE,
  layout = NA,
  restore.par = TRUE
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: porcupine "porc", cantábrico "cant", Golfo de Cádiz "arsa"}

\item{dns}{elique si los datos se toman del ordenador ("local") o del servidor ("serv")}

\item{incl0}{Si T se incluyen los lances nulos}

\item{xlims}{Define los limites longitudinales del mapa, si se deja en NA toma los límites de long del área definida en la campaña}

\item{ylims}{Define los limites latitudinales del mapa, si se deja en NA toma los límites de lat del área definida en la campaña}

\item{col}{Define el color de los segmentos}

\item{lwd}{Define el ancho de los segmentos}

\item{cexlab}{Tamaño de los números de lance utilizados en el gráfico de armap.lan parte del gráfico}

\item{places}{si T por defecto, incluye las etiquetas de países y ciudad en tierra, no funciona en Porcupine}

\item{Nlans}{se T incluye los números de lance por encima de los segmentos}

\item{bw}{Si T gráfico en blanco y negro por default, si F gráfico en color}

\item{rumbo}{si T incluye un punto al final del lance, en el punto de la virada, sólo en el segundo mapa}
}
\value{
Devuelve un mapa, pero no objetos de datos.
}
\description{
Muestra un mapa con el número de lance en el punto del lance y debajo o al lado mapa de los lances en segmentos
}
\examples{
\dontrun{
arMapLansGPS64("N20","cant","local",xlims=c(-9.8,-7.8),ylims=c(43.5,44.2))
}
}
\seealso{
\link{datlan.camp64}, \link{qcdistlan.camp64}

Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol64.camp}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{armap.camp64}()},
\code{\link{maparea64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()},
\code{\link{maphistal64}()}

Other PescaWin: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapLansGPS64}()}
}
\concept{PescaWin}
\concept{mapas}
