% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArteParComp64.R, R/ArteParCompC64.R
\name{ArteParComp64}
\alias{ArteParComp64}
\title{Comparación de parámetros del arte entre CAMP IEO y DATRAS}
\usage{
ArteParComp64(
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  var = c("calones", "puertas", "vertical"),
  incl2 = TRUE,
  es = TRUE,
  bw = TRUE,
  ti = TRUE,
  sub = TRUE,
  out.dat = FALSE,
  profrange = NA,
  Nlans = TRUE,
  lan.cex = 0.8,
  cex.leg = 1.1,
  graf = FALSE,
  xpng = 1200,
  ypng = 600,
  ppng = 15
)

ArteParComp64(
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  var = c("calones", "puertas", "vertical"),
  incl2 = TRUE,
  es = TRUE,
  bw = TRUE,
  ti = TRUE,
  sub = TRUE,
  out.dat = FALSE,
  profrange = NA,
  Nlans = TRUE,
  lan.cex = 0.8,
  cex.leg = 1.1,
  graf = FALSE,
  xpng = 1200,
  ypng = 600,
  ppng = 15
)
}
\arguments{
\item{camp}{Campaña: Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX", Arsa otoño "2XX"}

\item{zona}{Zona: "cant", "porc", "arsa"}

\item{dns}{Origen de datos: "local" o "serv"}

\item{var}{Variable a comparar: "calones"/"c" (wing spread), "puertas"/"p" (door spread) o "vertical"/"v" (net opening)}

\item{incl2}{Si TRUE incluye lances especiales (default TRUE)}

\item{es}{Si TRUE rótulos en español, si FALSE en inglés (default TRUE)}

\item{bw}{Si TRUE gráfico en blanco y negro, si FALSE en color (default TRUE)}

\item{ti}{Si TRUE añade título a cada panel (default TRUE)}

\item{sub}{Si TRUE añade subtítulo con la campaña y trimestre (default TRUE)}

\item{out.dat}{Si TRUE devuelve el merge de datos CAMP vs DATRAS (default FALSE)}

\item{profrange}{Rango de profundidad c(min,max) para filtrar; NA no filtra (default NA)}

\item{Nlans}{Si TRUE muestra números de lance en vez de puntos (default TRUE)}

\item{lan.cex}{Tamaño de las etiquetas de número de lance (default 0.8)}

\item{cex.leg}{Tamaño de letra de los ejes (default 1.1)}

\item{graf}{Si FALSE el gráfico va a pantalla; si es un string guarda como PNG con ese nombre}

\item{xpng}{Anchura del PNG en píxeles (default 1200)}

\item{ypng}{Altura del PNG en píxeles (default 600)}

\item{ppng}{Pointsize del PNG (default 15)}
}
\value{
Si out.dat=TRUE devuelve un data.frame con los datos CAMP e IEO fusionados por lance

Si out.dat=TRUE devuelve un data.frame con los datos CAMP e IEO fusionados por lance
}
\description{
Crea gráficos comparando los datos del comportamiento del arte en los lances de una campaña
con la profundidad, entre los datos del CAMP IEO y los datos de DATRAS.
Sustituye a ArteParCompC, ArteParCompD y ArteParCompV.

Crea gráficos comparando los datos del comportamiento del arte en los lances de una campaña
con la profundidad, entre los datos del CAMP IEO y los datos de DATRAS.
Sustituye a ArteParCompC, ArteParCompD y ArteParCompV.
}
\examples{
ArteParComp64("N23", zona="cant", var="calones", Nlans=FALSE)
ArteParComp64("N23", zona="cant", var="puertas", Nlans=TRUE, lan.cex=2)
ArteParComp64("P25", zona="porc", var="vertical")
ArteParComp64("N23", zona="cant", var="calones", Nlans=FALSE)
ArteParComp64("N23", zona="cant", var="puertas", Nlans=TRUE, lan.cex=2)
ArteParComp64("P25", zona="porc", var="vertical")
}
\seealso{
Other gear: 
\code{\link{ArtePars64}()}

Other gear: 
\code{\link{ArtePars64}()}
}
\concept{gear}
