% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscacod64.R
\name{buscacod64}
\alias{buscacod64}
\title{Función de búsqueda del código del grupo y la familia}
\usage{
buscacod64(nomb, zona = "cant", dns = c("local", "serv"))
}
\arguments{
\item{nomb}{Nombre científico de la especie o fragmento del nombre entre ""}
}
\description{
Busca el código del grupo, especie y la familia a partir del nombre o fragmento de la especie. Es independiente de mayúsculas y minúsculas y puede sacar más de una especie si una parte de sus nombres coincide.
}
\examples{
buscacod("sph")
}
\seealso{
Other datos_especies: 
\code{\link{BuscaAphia64}()},
\code{\link{buscaesp64}()},
\code{\link{camptoyear}()}
}
\concept{datos_especies}
