% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscaesp64.R
\name{buscaesp64}
\alias{buscaesp64}
\title{Función de búsqueda del nombre de una especie}
\usage{
buscaesp64(gr, esp, zona, dns = "local", id = "l")
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{id}{Salida como: "l" Nombre científico en latín, "e" nombre en español, "i" nombre en inglés "a" código APHIA}
}
\description{
Busca el nombre de una especie a partir del código del grupo y especie.Puede dar el nombre científico, común o en inglés
}
\examples{
buscaesp(1,50)
}
\seealso{
Other datos_especies: 
\code{\link{BuscaAphia64}()},
\code{\link{buscacod64}()},
\code{\link{camptoyear}()}
}
\concept{datos_especies}
