% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CampsDNS.camp64.R
\name{CampsDNS.camp64}
\alias{CampsDNS.camp64}
\title{Listado de campañas presentes en el directorio e información sobre ellas}
\usage{
CampsDNS.camp64(dns)
}
\arguments{
\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant, Golfo de Cádiz "Arsa", Mediterráneo "Medi"}
}
\value{
Un data.frame alineado con la lista de archivos presentes y el año
}
\description{
Función ROBUSTA de acceso a datos sin ODBC:
Alinea correctamente los archivos (CAMP, LANCE, FAUNA, ETC) usando el código de campaña (últimos 3 caracteres)
como clave única. Si falta un archivo (ej. no hay HIDRO), deja el hueco con un punto en lugar de descolocar la lista.
}
\examples{
CampsDNS.camp("Cant")
}
\concept{Series datos}
