% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CamptoHH64.R
\name{CAMPtoHH64}
\alias{CAMPtoHH64}
\title{Exporta datos de formato CAMP a formato DATRAs HH}
\usage{
CAMPtoHH64(camp, zona = "cant", dns = c("local", "serv"), quart = T, incl2 = F)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "Cant", Golfo de Cádiz "Arsa"}

\item{quart}{si F deja en cada lance el valor del trimestre en que se realizó el lance, si T se deja el que tiene la campaña por defecto, 1 para Arsa 1Q, 3 para Porcupine y 4 para Arsa 4Q y Demersales Northern Shelf}

\item{incl2}{Si F deja fuera los lances especiales que actualmente no se transmiten a DATRAS, si T los incluye}
}
\value{
Devuelve un data.table con datos de cada lance en el formato HH de DATRAS. DATRAS requiere que los datos no tengan cabecera y el trimestre sea el que corresponde a la campaña, además de no tener "". Por ello se debe pasar a fichero con la orden: write.table(CAMPtoHH(Xyy,dns),"nombrearchivo.csv",sep=",",quote=F,col.names=F,row.names=F))
}
\description{
Función de Salida de datos a DATRAS:
Extrae las características de los lances para una campaña y los transforma en formato DATRAS HH.
}
\examples{
# CAMPtoHH("P01","Porc")
}
\seealso{
{\code{CAMPtoHL}}
}
