% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecolgr.camp64.R
\name{ecolgr.camp64}
\alias{ecolgr.camp64}
\title{Índices ecológicos para un grupo en una campaña}
\usage{
ecolgr.camp64(
  gr,
  esp = 999,
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  ind = "n"
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros. 9 incluye todos los grupos a excepción del 6}

\item{esp}{ha de ser 999 cuando se quiere incluir todas las especies del grupo, o elegir todas las especies deseadas con los codigos de las especies}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{ind}{Elige el valor sobre el que se calculan los índices de diversidad, dominancia....}
}
\value{
Devuelve un data.frame con campos lan,lat,long,prof,div (Diversidad de Shanon),numbesp (riqueza: número de especies),simp (diversidad de simpson),domsimp (indice de dominancia de simpson).
}
\description{
Utiliza los datos del Camp para calcular la riqueza, diversidad y dominancia
}
\examples{
ecolgr.camp(1,999,"P08","Porc",ind="n")
ecolgr.camp(1,999,"M08","Medi",ind="n")
}
\seealso{
{\link{MapEcol.camp}}

Other ecologia: 
\code{\link{MapEcol.camp64}()},
\code{\link{ecolmatrix.camp64}()}
}
\concept{ecologia}
