% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICESarea64.R
\name{getICESarea64}
\alias{getICESarea64}
\title{StatRec y División ICES (Area) para una campaña — versión estricta sin helpers
Requisitos (en lances, todo en minúsculas):
latitud_l, latitud_v, longitud_l, longitud_v, prof_l, prof_v, validez
(y con redux=TRUE: lat, long, prof)
Requisitos (en Area): ICESNAME, Area}
\usage{
getICESarea64(
  camp,
  zona,
  dns = c("local", "serv"),
  incl2 = TRUE,
  incl0 = FALSE,
  verbose = FALSE
)
}
\description{
StatRec y División ICES (Area) para una campaña — versión estricta sin helpers
Requisitos (en lances, todo en minúsculas):
latitud_l, latitud_v, longitud_l, longitud_v, prof_l, prof_v, validez
(y con redux=TRUE: lat, long, prof)
Requisitos (en Area): ICESNAME, Area
}
