% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafhistbox64.R
\name{grafhistbox64}
\alias{grafhistbox64}
\title{Evolución en biomasa o abundancia de la especie en la serie histórica}
\usage{
grafhistbox64(
  gr,
  esp,
  camps,
  zona = "porc",
  dns = c("local", "serv"),
  ind = "p",
  cor.time = TRUE,
  kg = TRUE,
  ci.lev = 0.8,
  DLS = F,
  DLSrat = c(2, 5),
  idi = "l",
  SE = TRUE,
  es = TRUE,
  excl.sect = NA,
  sector = NA,
  ti = TRUE,
  Nas = FALSE,
  ymax = NA,
  mar = NA,
  tline = FALSE,
  years = TRUE,
  sub = FALSE,
  cex.leg = 1.1,
  nboot = 1000
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros. 9 incluye todos los grupos a excepción del 6}

\item{esp}{Codigo de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{campañas de la serie de datos a representar en el gráfico de abundancias Demersales Nsh, Porcupine Psh, Arsa primavera As1 y Arsa otoño As2}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew" o "Porc", Cantábrico "Cant" o "Cnew", Golfo de Cádiz "Arsa" y Mediterráneo "Medi" P}

\item{ind}{Parámetro a representar saca los datos en "p"eso o "n"úmero}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{kg}{Si T el gráfico está en kgs, si F en gramos}

\item{ci.lev}{El intervalo de confianza a representar}

\item{DLS}{si T muestra las medias de los dos últimos años y de los cinco precedentes}

\item{DLSrat}{ratio Elige el numero de años para el enfoque DLS 3/2 short life, 5/2 long life, vector con dos números, intervalo reciente intervalo histórico}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{SE}{Si T dibuja las cajas representando el error estándar, si F no aparecen la cajas}

\item{es}{Si T ejes y unidades en español, si F en inglés}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{sector}{Alternativa a excl.sect para un sólo sector. Si especificado como carácter solo toma el sector elegido}

\item{ti}{Si T muestra el título con el nombre de la especie (y el sector si se elige sect o excl.sect si falso no muestra nada. También se puede dar un valor y aparece el texto incluido o dar un list(label=(título), font, cex)}

\item{Nas}{Si F no representa las cajas en los sectores/estratos en los que algún sector o estrato tiene un sólo lance. Si T utiliza el valor del estrato y evita los NAs}

\item{ymax}{Valor máximo del eje y}

\item{mar}{Sirve para definir los margenes de la figura alrededor del gráfico mismo. Por defecto es NA y deja los definicos por defecto por R. Si se definen tiene que ser un vector con cuatro valores numéricos en líneas: c(bottom,left,upper,right)}

\item{tline}{Si T dibuja una línea de tendencia a traves de un glm con los datos de abundancia. gráficos evaluación MSFD.}

\item{years}{Si T saca los años como nombre de campaña en el eje de las equis en vez de de campañas}

\item{sub}{Añade un subtítulo con el valor que se le ponga si no F}

\item{cex.leg}{Sirve para modificar los tamaños de letra del gráfico, funciona con ejes y títulos. Por defecto en 1.1}

\item{nboot}{Número de réplicas bootstrap. Reducir (ej. 200) para exploración rápida, 1000 para resultados finales.}

\item{las}{Sirve para modificar la dirección de las etiquetas en el eje de las equis, por defecto F pero se cambia cuando se elige years=T, si no se cambia salen paralelas}
}
\value{
Crea una gráfica de evolución de las abundancias en biomasa o número y devuelve en consola un data.frame con columnas: avg,SE (error estándar),camp
}
\description{
Gráfico de cajas con la evolución en biomasa o abundancia en la serie histórica
}
\examples{
grafhistbox(1,45,Nsh[7:27],"Cant",es=FALSE,years=TRUE,tline=TRUE,DLS=FALSE,ti=TRUE,sub=TRUE)
grafhistbox(1,45,Nsh[1:31],"Cant",es=FALSE,years=TRUE,tline=FALSE,DLS=TRUE,ti=TRUE,sub=TRUE)
}
\seealso{
{\link{grafhistbox.comp}}
}
