% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maparea64.r
\name{maparea64}
\alias{maparea64}
\title{Mapa general del banco de Porcupine}
\usage{
maparea64(
  es = TRUE,
  leg = TRUE,
  bw = FALSE,
  dens = 0,
  strat = FALSE,
  places = TRUE,
  xlims = c(-15.5, -8.2),
  ylims = c(50.5, 54.5)
)
}
\arguments{
\item{es}{Textos en español (TRUE) o inglés (FALSE)}

\item{leg}{Incluye leyenda con los colores/patrones de los estratos}

\item{bw}{Si TRUE mapa en blanco y negro; si FALSE colorea estratos y sectores}

\item{dens}{Si > 0 las superficies de los estratos llevan patrón de líneas}

\item{strat}{Si TRUE dibuja contorno de estratos}

\item{places}{Si TRUE añade Galway y etiqueta IRLANDA/IRELAND}

\item{xlims, ylims}{Límites del mapa}
}
\description{
Mapa con la estratificación y el área general del banco de Porcupine incluyendo parte de Irlanda
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp64}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{armap.camp64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()},
\code{\link{maphistal64}()}

Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp64}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{armap.camp64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()},
\code{\link{maphistal64}()}
}
\concept{Porcupine}
\concept{Porcupine
Mapa general del banco de Porcupine

Mapa con la estratificación y el área general del banco de Porcupine
incluyendo parte de Irlanda.}
\concept{mapas}
