% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maparea64.r
\name{maparea64}
\alias{maparea64}
\title{Mapa general del banco de Porcupine}
\usage{
maparea64(es = TRUE, leg = TRUE, bw = FALSE, dens = 0, strat = FALSE)
}
\arguments{
\item{es}{Textos en español, si F en inglés}

\item{leg}{incluye la leyenda con los colores/patrones de los estratos}

\item{bw}{Si T mapa en blanco y negro, si F colorea los estratos y sectores}

\item{dens}{si mayor de 0 las superficies de los estratos tienen patrones de líneas}
}
\description{
Mapa con la estratificación y el área general del banco de Porcupine incluyendo parte de Irlanda
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp64}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{armap.tot64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()},
\code{\link{maphistal64}()}
}
\concept{Porcupine}
\concept{mapas}
