% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maphist64.R
\name{maphist64}
\alias{maphist64}
\title{Mapas de distribucion en varias campañas}
\usage{
maphist64(
  gr,
  esp,
  camps,
  zona = "cant",
  dns = "local",
  cor.time = TRUE,
  incl2 = TRUE,
  bw = TRUE,
  ti = TRUE,
  sub = NULL,
  plot = TRUE,
  out.dat = FALSE,
  ind = "p",
  idi = "l",
  layout = NA,
  leg = TRUE,
  pts = FALSE,
  ceros = TRUE,
  escmult = 0.25,
  cexleg = 1,
  years = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 desechos y otros, 9 escoge todos los orgánicos pero excluye desechos}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{Campaña a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o  "Pnew", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{incl2}{Si F no tiene en cuenta los lances especiales, si T si los tiene en cuenta, pero da problemas por que no puede calcular las abundancias estratificadas}

\item{bw}{Gráfico en blanco en negro si T o en color si F}

\item{ti}{Añade el nombre de la especie en latín sin T, si F no añade titulo}

\item{sub}{Añade un subtítulo debajo del gráfico, sin texto por defecto.}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{ind}{Parámetro a representar saca los datos en "p"eso o "n"úmero}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}

\item{leg}{Si T añade la leyenda}

\item{pts}{Saca los puntos de los datos}

\item{ceros}{Añade puntos para los datos igual a 0 si T, si F no incluye x en los valores 0}

\item{escmult}{Varía la relación de tamaño de los puntos con la leyenda y el máximo en los datos}

\item{cexleg}{Varía el tamaño de letra de los ejes y del número de la leyenda}

\item{years}{Si T saca los años como nombre de campaña en los paneles lattice de campañas}
}
\value{
Si out.dat=TRUE devuelve un data.frame con columnas: lan,lat,long,prof,peso.gr,numero (de individuos entre tmin y tmax),camp, si out.dat=F saca el gráfico en pantalla o como objeto para combinar con otros gráficos con print.trellis
}
\description{
Crea mapas con la distribución en biomasa o numero para distintas zonas: Porcupine (dns="Pnew"), el Cantábrico (dns=Cant), Cádiz= (dns=Arsa), y el Mediterráneo (dns=Medi)
}
\examples{
maphist(1,50,Nsh[7:27],"Cant",layout=c(3,7),years=TRUE)
maphist(1,50,As2[15:18],"Arsa",layout=c(2,2),years=TRUE,sub="ARSA 2nd quarter")
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapLansGPS64}()},
\code{\link{maphistal64}()}
}
\concept{mapas}
