% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maphistage64.r, R/maphistage64_1.r
\name{maphistage64}
\alias{maphistage64}
\title{Mapa distribución por edad en la campaña}
\usage{
maphistage64(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  age,
  plus = 8,
  cor.time = TRUE,
  n.ots = FALSE,
  AltAlk = NA,
  incl2 = TRUE,
  bw = TRUE,
  ti = TRUE,
  plot = TRUE,
  out.dat = FALSE,
  ind = "n",
  idi = "l",
  es = TRUE,
  layout = NA,
  ceros = FALSE,
  years = TRUE,
  mediahora = 1,
  lances = NULL,
  excl.sect = NA
)

maphistage64(
  gr,
  esp,
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  age,
  plus = 8,
  cor.time = TRUE,
  n.ots = FALSE,
  AltAlk = NA,
  incl2 = TRUE,
  bw = TRUE,
  ti = TRUE,
  plot = TRUE,
  out.dat = FALSE,
  ind = "n",
  idi = "l",
  es = TRUE,
  layout = NA,
  ceros = FALSE,
  years = TRUE,
  mediahora = 1,
  lances = NULL,
  excl.sect = NA
)
}
\arguments{
\item{gr}{Grupo de la especie: Solo hay dados de edad para algunos peces y cigala ?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña o campañas a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa".}

\item{dns}{elige si se trabaja con archivos del ordenador ("local") o del servidor ("serv")}

\item{age}{Edad solicitada}

\item{plus}{Edad plus: incluir la edad considerada como plus, solo afecta si se pide como plus la edad solicitada que suma todas las edades mayores}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{n.ots}{Interno para decir que en la clave no se saca el número de otolitos sino proporciones}

\item{AltAlk}{Clave talla edad alternativa sin ruta ni extensión, NA por defecto usa la clave de la campaña edadXYY.dbf}

\item{incl2}{si T representa los datos de lances especiales, si F los excluye}

\item{bw}{Gráfico en blanco y negro si T o en color si F}

\item{ti}{Si T añade titulo al mapa, el nombre de la especie en latín}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{out.dat}{Si T devuelve los datos como data.frame (compatible con plot=T o F)}

\item{ind}{Parámetro a representar saca los datos en "p"eso o "n"úmero}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{es}{Si T saca los ejes y leyendas en español, si F lo saca en inglés}

\item{layout}{Organización de gráficos en filas columnas c(r,c)}

\item{ceros}{Añade puntos para los datos igual a 0 si T, si F no incluye x en los valores 0}

\item{years}{Si T permite sacar los años como nombre de campaña en los paneles lattice de campañas}

\item{mediahora}{Valor para obtener abundancias por hora si media hora es mayor}

\item{lances}{Vector de números de lance a incluir tras el cálculo. NULL (por defecto) incluye todos.
Filtro post-hoc alternativo a excl.sect cuando se prefiere filtrar por número de lance.}

\item{excl.sect}{Sectores a excluir como carácter (ej. c("8C","8B") para quedarse solo con 9a).
NA por defecto no excluye ninguno. Se pasa directamente a datagegr.camp64 y datlan.camp64,
por lo que actúa sobre los lances antes del cálculo de abundancias por edad.
Combinado con AltAlk permite usar claves talla-edad específicas de cada división ICES.}
}
\value{
Si out.dat=TRUE devuelve un data.frame con columnas: lan,lat,long,0,1,2,...,Age+,camp.
Si out.dat=FALSE saca el gráfico en pantalla (plot=TRUE) o como objeto trellis (plot=FALSE)
para combinar con otros gráficos con print.trellis.

Si out.dat=TRUE devuelve un data.frame con columnas: lan,lat,long,0,1,2,...,Age+,camp.
Si out.dat=FALSE saca el gráfico en pantalla (plot=TRUE) o como objeto trellis (plot=FALSE)
para combinar con otros gráficos con print.trellis.
}
\description{
Mapa de la distribución geográfica por edad de la especie en las campañas solicitadas

Mapa de la distribución geográfica por edad de la especie en las campañas solicitadas
}
\examples{
maphistage64(1,42,c("N24","N25"),"cant","local",age=2)
maphistage64(1,42,"N25","cant","local",age=2,lances=45:67,AltAlk="edad942",out.dat=TRUE)
maphistage64(1,42,c("N24","N25"),"cant","local",age=2)
# Solo división 9a con clave específica:
maphistage64(1,42,"N25","cant","local",age=2,excl.sect=c("8C","8B"),AltAlk="edad942",out.dat=TRUE)
# Equivalente directo sin mapa:
datagegr.camp64(1,42,"N25","cant","local",excl.sect=c("8C","8B"),AltAlk="edad942")
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp64}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{armap.tot64}()},
\code{\link{armap64}()},
\code{\link{maparea64}()},
\code{\link{maphist64}()},
\code{\link{maphistal64}()}

Other edades: 
\code{\link{AbAgStatRec.camp64}()},
\code{\link{GetAlk.camp64}()},
\code{\link{bubbage.camp64}()},
\code{\link{datagegr.camp64}()},
\code{\link{edadsect.camp64}()},
\code{\link{edadstr.camp64}()},
\code{\link{grafedtal.camp64}()},
\code{\link{grafedtal.camps64}()},
\code{\link{logabage.camp64}()}

Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp64}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{NepFUs.camp64}()},
\code{\link{armap.tot64}()},
\code{\link{armap64}()},
\code{\link{maparea64}()},
\code{\link{maphist64}()},
\code{\link{maphistal64}()}

Other edades: 
\code{\link{AbAgStatRec.camp64}()},
\code{\link{GetAlk.camp64}()},
\code{\link{bubbage.camp64}()},
\code{\link{datagegr.camp64}()},
\code{\link{edadsect.camp64}()},
\code{\link{edadstr.camp64}()},
\code{\link{grafedtal.camp64}()},
\code{\link{grafedtal.camps64}()},
\code{\link{logabage.camp64}()}
}
\concept{edades}
\concept{mapas}
