% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maphistal64.r
\name{maphistal64}
\alias{maphistal64}
\title{Mapa distribución entre tallas tmin y tmax}
\usage{
maphistal64(
  gr,
  esp,
  camps,
  zona = "cant",
  dns = "local",
  tmin = 0,
  tmax = 999,
  cor.time = TRUE,
  incl2 = TRUE,
  ind = "n",
  sex = NA,
  bw = TRUE,
  ti = TRUE,
  sub = NULL,
  plot = TRUE,
  out.dat = FALSE,
  idi = "l",
  layout = NA,
  leg = TRUE,
  ceros = TRUE,
  escmult = 0.25,
  cexleg = 1,
  years = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustaceos 3 moluscos 4 equinodermos 5 invertebrados. 6 Desechos y otros inorgánicos no tiene sentido sacar tallas, sólo recogidas en peces, crustáceos decápodos y algunos moluscos}

\item{esp}{Código de la especie numerico o caracter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{zona}{Elige el origen de las bases de datos: Porcupine "porc", Cantábrico "cant", Golfo de Cádiz "arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{dns}{elige de dónde se toman los datos, si del ordenador "local" o del servidor "serv"}

\item{tmin}{Talla mínima del intervalo de tallas a incluir}

\item{tmax}{Talla máxima del intervalo de tallas a incluir}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{incl2}{Si F no presenta los lances especiales, si T si los tiene en cuenta, pero puede dar problemas por que no puede calcular las abundancias estratificadas}

\item{ind}{Permite elegir entre número "n" o peso "p" peso sólo funciona cuando existen a y b en especies.dbf y se elige sólo una especie}

\item{sex}{Permite elegir entre machos(1), hembras(2) o indeterminados(3), NA escoge sin tener en cuenta el sexo}

\item{bw}{gráfico en blanco en negro si T o en color si F}

\item{ti}{Si T añade titulo al mapa, el nombre de la especie en latín}

\item{sub}{Añade un subtítulo debajo del gráfico, sin texto por defecto.}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{out.dat}{Si T el resultado final de la funcion es la figura en pantalla, pero los datos en objeto}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}

\item{leg}{Si T añade la leyenda}

\item{ceros}{Añade puntos para los datos igual a 0 si T, si F no incluye x en los valores 0}

\item{escmult}{Varía la relación de tamaño de los puntos con la leyenda y el máximo en los datos}

\item{cexleg}{Varía el tamaño de letra de los ejes y del número de la leyenda}

\item{years}{Si T saca los años como nombre de campaña en los paneles lattice de campañas}
}
\value{
Si out.dat=TRUE devuelve un data.frame con columnas: lan,lat,long,prof,numero (de individuos entre tmin y tmax),camp, si out.dat=F saca el gráfico en pantalla o como objeto para combinar con otros gráficos con print.trellis
}
\description{
Saca mapas de distribución de una especie entre tallas tmin y tmax para varias campañas
}
\examples{
maphistal(1,50,Psh[1:12],"Porc",1,23,layout=c(4,3),out.dat=TRUE)
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp64}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{armap.tot64}()},
\code{\link{maparea64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()}
}
\concept{mapas}
\concept{tallas}
