% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapIberia64.R
\name{MapIberia64}
\alias{MapIberia64}
\title{Mapa de la Península Ibérica completo}
\usage{
MapIberia64(
  xlims = c(-9.72, 5),
  ylims = c(35.9, 44.5),
  lwdl = 1,
  cuadr = FALSE,
  cuadrMSFD = FALSE,
  ICESrect = FALSE,
  bw = F,
  ax = TRUE,
  wmf = FALSE,
  es = TRUE,
  places = TRUE,
  escmult = 1,
  new.dev = T
)
}
\arguments{
\item{xlims}{Define los limites longitudinales del mapa, los valores por defecto son los del total del ?rea de la campaña}

\item{ylims}{Define los limites latitudinales del mapa, los valores por defecto son los del total del ?rea de la campaña}

\item{lwdl}{Ancho de las líneas del mapa}

\item{cuadr}{Si T saca las cuadrículas de 5x5 millas naúticas}

\item{cuadrMSFD}{Si T dibuja caudr?cula de 10 millas naúticas utilizada para la evaluaci?n de la estrategia marina (MSFD)}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{bw}{si T mapa con tierra en gris, si F tierra en color}

\item{ax}{Si T saca los ejes x e y}

\item{wmf}{Si T saca a fichero metafile Iberia.emf}

\item{es}{Si T saca titulos y ejes en español}

\item{places}{Si T saca ciudades y puntos geográficos de referencia}

\item{escmult}{=1 aumenta o disminuye el tamaño de las etiquetas y textos menos de 1 lo disminuye, m?s de uno lo aumenta}
}
\value{
Saca en pantalla el mapa y es utilizada por otras funciones
}
\description{
Función auxiliar para sacar mapas de la campaña MEDITS
}
\examples{
MapIberia()
}
\seealso{
Other mapas base: 
\code{\link{MapMedit}()}

Other Medits: 
\code{\link{MapMedit}()}
}
\concept{Medits}
\concept{mapas base}
