% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NepFUs.camp64.R
\name{NepFUs.camp64}
\alias{NepFUs.camp64}
\title{Información anual sobre cigala por FUs en el Cantábrico y Galicia}
\usage{
NepFUs.camp64(
  camp = Nsh[length(Nsh)],
  zona = "cant",
  plot = TRUE,
  plotnep = TRUE,
  es = FALSE,
  ti = TRUE,
  ICESlab = TRUE,
  ICESrectcol = 2,
  ICESrect = TRUE,
  FU = NA,
  places = TRUE,
  country = TRUE,
  out.dat = TRUE
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{plot}{Saca el gráfico (T) o lo omite para dejar sólo los datos (F)}

\item{plotnep}{si T presenta todos los lances en los que ha habido cigala en la campaña, además de los lances en cada FU, con o sín captura. Si F sólo saca los lances en cada FU sin marcar si ha habido cigala o no.}

\item{es}{si T letreros en español, si F en inglés (F por defecto)}

\item{ICESlab}{Si T incluye las etiquetas de los rectángulos ICES}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{FU}{pinta una o varias unidades funcionales, a elegir FU26, FU25 o FU31 con grueso lwd=2 y color rojo}

\item{places}{Si T saca etiquetas de principales ciudades en el mapa, si F se omiten los letreros}

\item{country}{si T saca el país}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos de abundancia de cigala en cada FU con datos de biomasa, numero, desviaciones estándar y número de lances en cada FU.}

\item{dns}{Sólo disponible para el Cantábrico "Cant", Golfo de Cádiz "Arsa", combinados con "dnsred" busca los datos en el servidor de Santander si se han creado las odbcs}

\item{year}{si T incluye una columna con el año al final de los datos}

\item{icesrectcol}{Color para los rectángulos ICES}

\item{ICESlabcex}{tamaño del ICESlab en cex, .5 por defecto subirlo si se quiere más grande}
}
\value{
Produce un gráfico con los lances en los que ha habido cigala en el lance y especialmente los lances en cada FU dentro de Demersales FU25,FU26 y FU31
}
\description{
Función de acceso a datos:
Extrae las características de los lances para una campaña determinada y las capturas de cigala por unidad funcional
}
\details{
Un problema que ocurre al utilizar el CampR con ficheros dbf de las primeras campañas
puede ser que al fichero lanceXXX.dbf le falte algún campo, habitualmente
el campo \strong{ESTN} utilizado en las últimas versiones del \strong{CAMP} para ligar lances con las estaciones de CTD.
El error usual es \strong{$ operator is invalid for atomic vectors}
Si se detecta este error revisar la estructura de lanceXXX.dbf con la de
otros ficheros de lances de los últimos años
}
\examples{
  NepFUs.camp(camp="N20",dns="Cant",plot=T,plotnep=T)
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp64}()},
\code{\link{MapLansGPS64}()},
\code{\link{MatchHidro.ctd64}()},
\code{\link{armap.tot64}()},
\code{\link{armap64}()},
\code{\link{maparea64}()},
\code{\link{maphist64}()},
\code{\link{maphistage64}()},
\code{\link{maphistal64}()}
}
\concept{mapas}
