% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcdistlan.camp64.r
\name{qcdistlan.camp64}
\alias{qcdistlan.camp64}
\title{Comprueba la distancia recorrida en los lances y la consistencia con recorrido y la velocidad en los datos del CAMP}
\usage{
qcdistlan.camp64(
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  todos = FALSE,
  pc.error = 2
)
}
\arguments{
\item{camp}{campaña a revisar los datos en formato Camp Xyy}

\item{zona}{Origen de bases de datos: "cant" cantábrico, "porc"  Porcupine, "arsa" para el Golfo de Cádiz y "medi" para MEDITS}

\item{todos}{Por defecto F. Si T lista todos los lances con valores, si no sólo los que pc.error>error}

\item{pc.error}{porcentaje de error aceptable para no mostrar los lances como erróneos}
}
\value{
Devuelve un data.frame con campaña, lance, recorrido, recorrido según la fórmula de Haversine, recorrido según la velocidad x el tiempo, velocidad, tiempo, rumbo, rumbo estimado según posiciones,velocidad calculada a partir de la distancia y el tiempo, y los porcentajes de errores de distancia, velocidad y rumbo.
}
\description{
Sirve para control de calidad y asegurarse que los datos de distancias y posiciones son correctos
}
\examples{
qcdistlan.camp64("C14","cant","local",pc.error=.01)
qcdistlan.camp64("216","arsa","local",pc.error=.01)
}
\references{
distHaversine function gives the haversine calculation of distance between two geographic points \code{\link[geosphere]{distHaversine}}
}
\seealso{
{\link{MapLansGPS}}

Other Control de calidad: 
\code{\link{qcLW64.camp}()},
\code{\link{qcTalPez64.camp}()},
\code{\link{qcTalPez64.lan}()},
\code{\link{qcdtall64.camp}()},
\code{\link{qcdtallrev64.camp}()},
\code{\link{qclandatr.camp64}()}
}
\concept{Control de calidad}
