% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcdtall64.camp.R
\name{qcdtall64.camp}
\alias{qcdtall64.camp}
\title{Revisión del histograma de distribución de tallas}
\usage{
qcdtall64.camp(
  gr,
  esp,
  camp = "P12",
  zona = "cant",
  dns = c("local", "serv"),
  lance,
  ti = FALSE,
  legend = TRUE,
  idi = "l",
  ymax = NA
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos, 3 moluscos, 4 equinodermos, 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios}

\item{camp}{Campaña: Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX", Arsa otoño "2XX"}

\item{zona}{Zona de las bases de datos: "cant", "porc", "arsa"}

\item{dns}{Origen de datos: "local" o "serv"}

\item{lance}{Número de lance a revisar}

\item{ti}{Si TRUE añade título al gráfico con el nombre de la especie en latín}

\item{legend}{Si TRUE añade leyenda de categorías}

\item{idi}{Nombre científico ("l") o nombre común ("e")}

\item{ymax}{Valor máximo del eje y (NA = automático)}
}
\value{
Gráfico con el histograma de individuos medidos divididos en categorías y ponderaciones
}
\description{
Revisa el histograma de distribución de tallas total de la especie en un lance concreto de una campaña.
Muestra la distribución de talla de cada categoría, los factores de ponderación por categoría,
el número total de individuos medidos y el número ideal de individuos a medir
(10 por talla: Gerritsen & McGrath 2007)
}
\examples{
qcdtall.camp64(1, 43, "P08", zona="porc", dns="local", lance=35)
}
\references{
Gerritsen & McGrath (2007). Precision estimates and suggested sample sizes for
length-frequency data. Fish. Bull. 106: 116-120.
}
\seealso{
\code{\link{dtall.camp64}}

Other Control de calidad: 
\code{\link{qcLW64.camp}()},
\code{\link{qcTalPez64.camp}()},
\code{\link{qcTalPez64.lan}()},
\code{\link{qcdistlan.camp64}()},
\code{\link{qcdtallrev64.camp}()},
\code{\link{qclandatr.camp64}()}
}
\concept{Control de calidad}
