% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcLW64.camp.R
\name{qcLW64.camp}
\alias{qcLW64.camp}
\title{Outliers capturas según tallas medidas y relación talla-peso}
\usage{
qcLW64.camp(
  gr,
  esp,
  camp = "P11",
  zona = "porc",
  dns = c("local", "serv"),
  margerr = 20,
  out.dat = FALSE,
  mm = FALSE,
  areg = NA,
  breg = NA
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos, 3 moluscos, 4 equinodermos, 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter}

\item{camp}{Campaña: Demersales "NXX", Porcupine "PXX", Arsa "1XX"/"2XX"}

\item{zona}{Zona: "cant", "porc", "arsa"}

\item{dns}{Origen de datos: "local" o "serv"}

\item{margerr}{Margen de error que delimita los datos dentro o fuera del intervalo (default 20)}

\item{out.dat}{Si TRUE devuelve los datos estimados, observados y el error (default FALSE)}

\item{mm}{Si TRUE la especie está medida en milímetros (default FALSE)}

\item{areg}{Coeficiente "a" alternativo de la regresión talla-peso (NA = usar el del CAMP)}

\item{breg}{Coeficiente "b" alternativo de la regresión talla-peso (NA = usar el del CAMP)}
}
\value{
Gráfico de errores por lance y tabla con los lances fuera del margen de error
}
\description{
Busca outliers en las capturas a partir de la relación talla-peso y los pesos y tallas
muestreadas. El tamaño de los círculos refleja el número de peces medidos y los distintos
colores marcan las distintas categorías si existe muestreo por categorías.
}
\examples{
qcLW.camp64(1, 50, "N25", zona="porc", dns="local")
}
\seealso{
Other Control de calidad: 
\code{\link{qcTalPez64.camp}()},
\code{\link{qcTalPez64.lan}()},
\code{\link{qcdistlan.camp64}()},
\code{\link{qcdtall64.camp}()},
\code{\link{qcdtallrev64.camp}()},
\code{\link{qclandatr.camp64}()}
}
\concept{Control de calidad}
