% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcTalPez64.camp.R
\name{qcTalPez64.camp}
\alias{qcTalPez64.camp}
\title{Comprueba que están medidas todas las especies en capturas y viceversa en la campaña}
\usage{
qcTalPez64.camp(
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  gr = 1,
  verbose = FALSE
)
}
\arguments{
\item{camp}{Campaña en formato Camp Xyy}

\item{zona}{Zona: "cant", "porc", "arsa"}

\item{dns}{Origen de datos: "local" o "serv"}

\item{gr}{Grupo: 1 peces, 2 crustáceos, 3 cefalópodos (default 1)}

\item{verbose}{Si TRUE muestra también los lances sin errores (default FALSE)}
}
\value{
Imprime en consola las discordancias encontradas lance por lance
}
\description{
Recorre todos los lances de una campaña llamando a \code{\link{qcTalPez.lan64}} en cada uno.
}
\examples{
qcTalPez64.camp("N25", zona="cant", dns="local", gr=1)
}
\seealso{
\code{\link{qcTalPez.lan64}}

Other Control de calidad: 
\code{\link{qcLW64.camp}()},
\code{\link{qcTalPez64.lan}()},
\code{\link{qcdistlan.camp64}()},
\code{\link{qcdtall64.camp}()},
\code{\link{qcdtallrev64.camp}()},
\code{\link{qclandatr.camp64}()}
}
\concept{Control de calidad}
