% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcTalPez64.lan.R
\name{qcTalPez64.lan}
\alias{qcTalPez64.lan}
\title{Comprueba concordancia entre capturas y tallas en un lance}
\usage{
qcTalPez64.lan(
  camp,
  zona = "cant",
  dns = c("local", "serv"),
  lan,
  gr = 1,
  verbose = FALSE
)
}
\arguments{
\item{camp}{Campaña en formato Camp Xyy}

\item{zona}{Zona: "cant", "porc", "arsa"}

\item{dns}{Origen de datos: "local" o "serv"}

\item{lan}{Lance a comprobar}

\item{gr}{Grupo: 1 peces, 2 crustáceos, 3 cefalópodos (default 1)}

\item{verbose}{Si TRUE muestra también los lances sin errores (default FALSE)}
}
\value{
Imprime en consola las discordancias encontradas
}
\description{
Comprueba coherencia entre faunaXXX.dbf y NtallXXX.dbf para un lance concreto:
especies capturadas sin medir, especies medidas sin captura, y discordancias
entre pesos y números del fauna vs los calculados desde las tallas.
}
\examples{
qcTalPez64.lan("N25", zona="cant", dns="local", lan=23, gr=1)
}
\seealso{
\code{\link{qcTalPez.camp64}}

Other Control de calidad: 
\code{\link{qcLW64.camp}()},
\code{\link{qcTalPez64.camp}()},
\code{\link{qcdistlan.camp64}()},
\code{\link{qcdtall64.camp}()},
\code{\link{qcdtallrev64.camp}()},
\code{\link{qclandatr.camp64}()}
}
\concept{Control de calidad}
